refactor: 메인 페이지 컴포넌트화 fix: Image 에러 해결
