+++ src/components/main/rankingImg.tsx
@@ -1,0 +1,146 @@
+export default function Ranking() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="439"
+      height="300"
+      fill="none"
+    >
+      <rect width="437.988" height="300" fill="#fff" rx="8" />
+      <path
+        stroke="#C4C4C4"
+        d="M.7891 97.0488H438.777M.7891 97.0488H438.777M.7891 97.0488H438.777M.7891 97.0488H438.777M.7891 97.0488H438.777"
+      />
+      <path
+        fill="#000"
+        d="M29.18 55.1304V72.585h-3.6903V58.6333h-.1023l-3.9971 2.5056v-3.2727l4.321-2.7358H29.18ZM123.45 66.214c1.437 0 2.548.1727 3.332.518.784.3453 1.176.8447 1.176 1.498 0 .6627-.392 1.162-1.176 1.498-.784.3453-1.895.518-3.332.518-1.428 0-2.534-.1727-3.318-.518-.784-.336-1.176-.8353-1.176-1.498 0-.6533.392-1.1527 1.176-1.498.784-.3453 1.89-.518 3.318-.518Zm.014 1.344c-.906 0-1.568.0513-1.988.154-.42.1027-.63.28-.63.532 0 .2333.21.406.63.518.42.1027 1.082.154 1.988.154.896 0 1.549-.0513 1.96-.154.42-.112.63-.2847.63-.518 0-.252-.21-.4293-.63-.532-.411-.1027-1.064-.154-1.96-.154Zm-5.292-9.352h10.57v1.4h-10.57v-1.4Zm5.306 1.736c1.4 0 2.473.1587 3.22.476.746.308 1.12.77 1.12 1.386 0 .6067-.374 1.0687-1.12 1.386-.747.308-1.82.462-3.22.462-1.382 0-2.45-.154-3.206-.462-.756-.3173-1.134-.7793-1.134-1.386 0-.616.378-1.078 1.134-1.386.756-.3173 1.824-.476 3.206-.476Zm0 1.246c-.794 0-1.382.0513-1.764.154-.383.0933-.574.2473-.574.462 0 .2053.191.3547.574.448.382.0933.97.14 1.764.14.812 0 1.404-.0467 1.778-.14.382-.0933.574-.2427.574-.448 0-.2147-.192-.3687-.574-.462-.374-.1027-.966-.154-1.778-.154Zm-.924-4.018h1.848v2.002h-1.848V57.17Zm-4.942 7.112h11.732v1.442h-11.732v-1.442Zm4.942-1.218h1.848v1.82h-1.848v-1.82Zm16.855-5.796h1.862v6.51h-1.862v-6.51Zm-4.074.49h2.044c0 1.176-.22 2.184-.658 3.024-.439.84-1.106 1.526-2.002 2.058-.887.532-2.016.938-3.388 1.218l-.644-1.498c.868-.1587 1.596-.3593 2.184-.602.597-.2427 1.073-.5273 1.428-.854.364-.3267.625-.6813.784-1.064.168-.3827.252-.7887.252-1.218v-1.064Zm-3.962 0h5.264v1.47h-5.264v-1.47Zm1.26 6.552h8.638v3.528h-6.762v1.4h-1.862v-2.744h6.748v-.756h-6.762V64.31Zm.014 4.368h8.946v1.442h-8.946v-1.442Zm10.751-5.306h11.718v1.47h-11.718v-1.47Zm4.928-1.946h1.862v2.52h-1.862v-2.52Zm-3.514-.588h8.974v1.456h-8.974v-1.456Zm0-3.01h8.904v1.47h-7.056v2.464h-1.848v-3.934Zm4.396 7.658c1.4 0 2.501.21 3.304.63.802.4107 1.204.9987 1.204 1.764 0 .756-.402 1.3393-1.204 1.75-.803.42-1.904.63-3.304.63s-2.502-.21-3.304-.63c-.794-.4107-1.19-.994-1.19-1.75 0-.7653.396-1.3533 1.19-1.764.802-.42 1.904-.63 3.304-.63Zm0 1.414c-.579 0-1.064.0373-1.456.112-.383.0653-.672.1727-.868.322-.196.14-.294.3173-.294.532 0 .224.098.4107.294.56.196.14.485.2473.868.322.392.0653.877.098 1.456.098.578 0 1.059-.0327 1.442-.098.392-.0747.686-.182.882-.322.196-.1493.294-.336.294-.56 0-.2147-.098-.392-.294-.532-.196-.1493-.49-.2567-.882-.322-.383-.0747-.864-.112-1.442-.112ZM226.539 58.3807v10.1818h-2.152V60.424h-.06l-2.332 1.4617v-1.9091l2.521-1.5959h2.023Zm6.413 10.4055c-.856-.0033-1.591-.2138-2.208-.6314-.613-.4176-1.085-1.0225-1.417-1.8146-.328-.7921-.49-1.745-.487-2.8587 0-1.1103.164-2.0565.492-2.8387.332-.7822.804-1.3772 1.417-1.7848.617-.411 1.351-.6165 2.203-.6165.851 0 1.584.2055 2.197.6165.616.4109 1.09 1.0075 1.422 1.7897.331.7789.495 1.7235.492 2.8338 0 1.117-.166 2.0715-.497 2.8637-.328.7921-.799 1.397-1.412 1.8146-.613.4176-1.347.6264-2.202.6264Zm0-1.7848c.583 0 1.049-.2933 1.397-.88.348-.5866.52-1.4666.517-2.6399 0-.7722-.08-1.4152-.239-1.9289-.156-.5138-.378-.8999-.666-1.1584-.285-.2585-.622-.3878-1.009-.3878-.581 0-1.045.29-1.393.87-.348.58-.523 1.4484-.527 2.6051 0 .7822.078 1.4352.234 1.9588.159.5204.383.9115.671 1.1733.289.2586.627.3878 1.015.3878Zm9.638 1.7848c-.855-.0033-1.591-.2138-2.207-.6314-.613-.4176-1.086-1.0225-1.417-1.8146-.328-.7921-.491-1.745-.487-2.8587 0-1.1103.164-2.0565.492-2.8387.331-.7822.804-1.3772 1.417-1.7848.616-.411 1.35-.6165 2.202-.6165.852 0 1.584.2055 2.198.6165.616.4109 1.09 1.0075 1.422 1.7897.331.7789.495 1.7235.492 2.8338 0 1.117-.166 2.0715-.497 2.8637-.329.7921-.799 1.397-1.412 1.8146-.614.4176-1.348.6264-2.203.6264Zm0-1.7848c.584 0 1.049-.2933 1.397-.88.348-.5866.521-1.4666.517-2.6399 0-.7722-.079-1.4152-.238-1.9289-.156-.5138-.378-.8999-.667-1.1584-.285-.2585-.621-.3878-1.009-.3878-.58 0-1.044.29-1.392.87-.348.58-.524 1.4484-.527 2.6051 0 .7822.078 1.4352.234 1.9588.159.5204.383.9115.671 1.1733.288.2586.626.3878 1.014.3878ZM306.053 58.9754v10.1818H303.9v-8.1385h-.059l-2.332 1.4617v-1.9091l2.521-1.5959h2.023ZM354.322 67v-1.5511l3.625-3.3559c.308-.2983.566-.5667.775-.8054.212-.2386.373-.4723.482-.701.11-.232.165-.4822.165-.7507 0-.2983-.068-.5551-.204-.7706-.136-.2187-.322-.3861-.557-.5021-.235-.1193-.502-.179-.801-.179-.311 0-.583.063-.815.189-.232.1259-.411.3065-.537.5419-.126.2353-.189.5153-.189.8402h-2.043c0-.6662.151-1.2446.452-1.7351.302-.4906.725-.8701 1.268-1.1385.544-.2685 1.17-.4027 1.879-.4027.73 0 1.364.1292 1.904.3878.544.2552.967.6098 1.268 1.0639.302.4541.453.9744.453 1.5611 0 .3844-.077.7639-.229 1.1385-.149.3745-.416.7904-.8 1.2478-.385.4541-.927.9993-1.626 1.6357l-1.487 1.4567v.0696h4.276V67h-7.259Zm12.811.2237c-.855-.0033-1.591-.2138-2.207-.6314-.614-.4176-1.086-1.0225-1.417-1.8146-.329-.7921-.491-1.745-.488-2.8587 0-1.1103.164-2.0565.493-2.8387.331-.7822.803-1.3772 1.416-1.7848.617-.411 1.351-.6165 2.203-.6165.852 0 1.584.2055 2.197.6165.617.4109 1.091 1.0075 1.422 1.7897.332.7789.496 1.7235.492 2.8338 0 1.117-.165 2.0715-.497 2.8637-.328.7921-.799 1.397-1.412 1.8146-.613.4176-1.347.6264-2.202.6264Zm0-1.7848c.583 0 1.049-.2933 1.397-.88.348-.5866.52-1.4666.517-2.6399 0-.7722-.08-1.4152-.239-1.9289-.155-.5138-.378-.8999-.666-1.1584-.285-.2585-.621-.3878-1.009-.3878-.58 0-1.044.29-1.392.87-.348.58-.524 1.4484-.527 2.6051 0 .7822.078 1.4352.234 1.9588.159.5204.382.9115.671 1.1733.288.2586.626.3878 1.014.3878ZM372.779 67v-1.5511l3.625-3.3559c.308-.2983.566-.5667.775-.8054.212-.2386.373-.4723.482-.701.11-.232.165-.4822.165-.7507 0-.2983-.068-.5551-.204-.7706-.136-.2187-.322-.3861-.557-.5021-.235-.1193-.502-.179-.801-.179-.311 0-.583.063-.815.189-.232.1259-.411.3065-.537.5419-.126.2353-.189.5153-.189.8402h-2.043c0-.6662.151-1.2446.452-1.7351.302-.4906.725-.8701 1.268-1.1385.544-.2685 1.17-.4027 1.879-.4027.73 0 1.364.1292 1.905.3878.543.2552.966.6098 1.267 1.0639.302.4541.453.9744.453 1.5611 0 .3844-.077.7639-.229 1.1385-.149.3745-.416.7904-.8 1.2478-.385.4541-.927.9993-1.626 1.6357l-1.487 1.4567v.0696h4.276V67h-7.259Zm12.537.1392c-.702 0-1.329-.1293-1.879-.3878-.547-.2585-.981-.6148-1.302-1.0689-.322-.454-.489-.9744-.502-1.5611h2.088c.023.3945.189.7143.497.9596.308.2452.674.3679 1.098.3679.339 0 .637-.0746.895-.2238.262-.1524.466-.3629.612-.6313.149-.2718.224-.5834.224-.9347 0-.358-.077-.6728-.229-.9446-.149-.2718-.356-.4839-.622-.6364-.265-.1524-.568-.2303-.909-.2336-.299 0-.589.0613-.87.1839-.279.1226-.496.29-.652.5021l-1.914-.343.483-5.3693h6.224v1.7599h-4.45l-.263 2.5505h.06c.179-.2519.449-.4607.81-.6265.361-.1657.766-.2485 1.213-.2485.613 0 1.16.1441 1.641.4325.48.2883.86.6844 1.138 1.1882.279.5005.416 1.0772.413 1.7301.003.6861-.156 1.2959-.478 1.8295-.318.5304-.763.948-1.337 1.2529-.57.3016-1.233.4524-1.989.4524Zm6.235-.0099c-.328 0-.61-.116-.845-.3481-.232-.2353-.348-.517-.348-.8451 0-.3248.116-.6032.348-.8352.235-.2321.517-.3481.845-.3481.318 0 .596.116.835.3481.239.232.358.5104.358.8352 0 .2187-.056.4193-.169.6015-.109.179-.254.3232-.433.4326-.178.106-.376.1591-.591.1591Zm6.494.0944c-.855-.0033-1.591-.2138-2.207-.6314-.614-.4176-1.086-1.0225-1.417-1.8146-.328-.7921-.491-1.745-.487-2.8587 0-1.1103.164-2.0565.492-2.8387.331-.7822.803-1.3772 1.417-1.7848.616-.411 1.35-.6165 2.202-.6165.852 0 1.584.2055 2.197.6165.617.4109 1.091 1.0075 1.422 1.7897.332.7789.496 1.7235.493 2.8338 0 1.117-.166 2.0715-.498 2.8637-.328.7921-.798 1.397-1.412 1.8146-.613.4176-1.347.6264-2.202.6264Zm0-1.7848c.583 0 1.049-.2933 1.397-.88.348-.5866.52-1.4666.517-2.6399 0-.7722-.079-1.4152-.239-1.9289-.155-.5138-.377-.8999-.666-1.1584-.285-.2585-.621-.3878-1.009-.3878-.58 0-1.044.29-1.392.87-.348.58-.524 1.4484-.527 2.6051 0 .7822.078 1.4352.234 1.9588.159.5204.382.9115.671 1.1733.288.2586.626.3878 1.014.3878Zm10.076-8.6207V67h-2.153v-8.1385h-.059l-2.332 1.4617v-1.9091l2.521-1.5959h2.023Zm3.131 10.3111c-.328 0-.61-.116-.845-.3481-.232-.2353-.348-.517-.348-.8451 0-.3248.116-.6032.348-.8352.235-.2321.517-.3481.845-.3481.318 0 .597.116.835.3481.239.232.358.5104.358.8352 0 .2187-.056.4193-.169.6015-.109.179-.253.3232-.432.4326-.179.106-.377.1591-.592.1591Zm6.07-10.3111V67h-2.152v-8.1385h-.06l-2.332 1.4617v-1.9091l2.521-1.5959h2.023ZM420.145 67l4.221-8.3523v-.0696h-4.917v-1.7599h7.144v1.7848L422.367 67h-2.222Z"
+      />
+      <g fill="red" clipPath="url(#a)">
+        <path d="M61.025 39.7769c-1.1277.2861-2.1884 1.089-2.6949 2.0211-.2485.4429-3.3638 8.8871-3.6601 9.9207-.3631 1.2274-.4396 2.261-.4396 5.8325v3.4884h-.3631c-.8792 0-1.9399.5722-2.3031 1.2367-.1051.1753-.2102.2399-.4204.2399-.8506 0-1.9113.7568-2.1502 1.532-.086.286-.1147 2.667-.1147 9.7638v9.3763h1.529V64.445l.2389-.2307c.3441-.3323.5256-.2584.6881.2768.2867.9321 1.4334 1.7166 2.5324 1.7166h.3631v2.1041c0 2.3348.0765 2.787.5925 3.7283.4014.7291 1.3093 1.5412 2.112 1.8735.6307.2676.6403.2768.7932.7844.086.2768.344.9874.5829 1.5689 1.099 2.7039 2.7235 5.1126 4.6348 6.8845 7.0621 6.5431 16.7427 4.2175 21.6068-5.1865.2389-.4522.4492-.9044.4683-.9874.0191-.1292-.1338-.2308-.6307-.4338-.3632-.1476-.669-.2676-.6881-.2492-.0095.0093-.2962.5538-.6402 1.2182-2.284 4.3559-5.6861 7.0784-9.6615 7.7521-2.4655.4152-5.2273-.1938-7.4635-1.6427-3.985-2.5841-6.7658-7.4106-7.6546-13.3077-.2389-1.5596-.2389-5.2049-.0095-6.6908.2389-1.4673.602-3.0085.9843-4.1621.1815-.5352.3249-1.0151.3249-1.0613 0-.0461-.3058-.1938-.6785-.3322l-.6785-.2399-.1338.323c-.4396 1.0059-1.1563 3.9314-1.3761 5.5648-.2867 2.1872-.2103 5.3526.1816 7.5675.0955.5353.1624.9875.1433 1.0059-.0764.0831-.7836-.729-1.0034-1.1536l-.2294-.4522v-5.6294c0-5.9894.0191-6.3032.4874-8.112.8505-3.313 2.6758-6.3585 5.1413-8.6103 2.3222-2.1226 4.7686-3.3961 7.7502-4.0144 1.3761-.2861 3.8894-.2861 5.2655 0 1.7297.3599 3.5167 1.0428 4.7973 1.8272.2771.1754.516.3138.5351.3138.0191 0 .2198-.2768.4492-.6091.3345-.4983.3822-.6183.2676-.6921a38.0719 38.0719 0 0 0-.9843-.563l-.841-.4799.0478-.5168c.086-.932.6594-1.7442 1.5481-2.1964.6594-.3414 1.7966-.3507 2.4369-.0277.5542.2769 1.0225.7014 1.2518 1.1444.1052.1938.927 2.3533 1.8157 4.7989 1.4144 3.8668 1.9878 5.6294 1.9973 6.1555 0 .0738-.2484-.3415-.5543-.9229-.6784-1.3012-1.3761-2.3441-2.3795-3.553-.8123-.969-2.4655-2.5748-2.5993-2.5194-.0478.0184-.2771.2584-.516.5352l-.4396.5076 1.1276 1.1074c.6116.6091 1.3188 1.3566 1.5577 1.6612 2.1406 2.7409 3.4498 5.9248 3.8034 9.3209.0669.6645.0956 2.7778.0765 6.0447l-.0287 5.0296-.3249.5814c-.2772.4892-.8314 1.1075-.9174 1.0152-.0096-.0185.0573-.4707.1529-1.0059 1.3187-7.4568-.8983-15.3749-5.6287-20.1369-3.6696-3.673-8.4669-4.9834-12.9679-3.5345-1.9495.6367-3.4881 1.5596-5.1509 3.1008-1.1276 1.052-2.3317 2.464-2.1979 2.5748.2771.2307 1.0798.7475 1.1467.7383.0478 0 .4301-.4246.8601-.9414 2.7809-3.3961 6.7468-5.0572 10.598-4.4482 3.2969.5168 6.2498 2.5379 8.5146 5.8141 3.4594 5.0296 4.568 12.2002 2.8669 18.6418-.1338.5076-.3058 1.0982-.3822 1.3104-.086.2308-.1051.4153-.0574.443.1529.0923 1.1659.4153 1.2901.4153.0669 0 .1912-.2399.2772-.5353.1529-.526.1529-.526.7836-.7936 1.2805-.5445 2.3604-1.855 2.6184-3.1839.0574-.2676.0956-1.4027.0956-2.5102v-2.021h.3631c1.099 0 2.2458-.7845 2.5325-1.7166.1624-.5352.344-.6091.688-.2768l.2389.2307v18.7433h1.529V73.812c0-6.986-.0286-9.4778-.1146-9.7454-.2485-.8213-1.2615-1.5504-2.1311-1.5504-.2294 0-.3345-.0554-.4396-.2399-.3631-.6645-1.4239-1.2367-2.3031-1.2367h-.3631v-3.4884c0-3.5715-.0765-4.6051-.4396-5.8325-.2962-1.0336-3.4116-9.4778-3.6601-9.9207-.516-.9506-1.5768-1.735-2.7331-2.0303-2.3508-.5907-4.8068.9044-5.3037 3.2392-.0956.4153-.1434.4984-.2676.4522-.4683-.1753-1.8539-.526-2.628-.6644-1.2041-.2123-3.899-.2215-5.084-.0093-.7549.1385-2.1788.4984-2.6471.6737-.1242.0462-.172-.0369-.2675-.4522-.5065-2.3441-2.9816-3.8483-5.342-3.23Zm2.2839 1.6058c.9557.4614 1.5482 1.3197 1.5673 2.2425l.0095.5445-.8983.4984c-1.3761.7659-2.3317 1.4673-3.5645 2.6117-1.7488 1.6242-3.0006 3.2484-4.0614 5.2603-.2771.526-.5065.9228-.5065.8767 0-.0461.086-.4338.1816-.8675.1911-.8121 3.2969-9.4132 3.6314-10.0315.3058-.5814.9652-1.089 1.7297-1.3289.4205-.1385 1.4525-.0369 1.9112.1938Zm-9.0785 22.2502v1.1259l-.4395-.0554c-.6594-.0646-1.0417-.4707-1.0417-1.0705 0-.6737.4492-1.0982 1.1754-1.1075l.3058-.0092v1.1167Zm36.3523-.7568c.2484.2492.2962.3599.2962.7475 0 .6091-.3822 1.0152-1.0416 1.0798l-.4396.0554V62.4885l.4396.0554c.3249.0277.5256.12.7454.3322Z" />
+        <path d="M70.86 50.1128c-2.1693.7568-2.8 3.3408-1.1946 4.882.6308.6091 1.1946.8306 2.1502.8213.6881 0 .8887-.0369 1.2997-.2399.7167-.3599 1.0225-.646 1.3856-1.292.2867-.5261.3154-.6553.3249-1.3659 0-.9228-.2293-1.4673-.86-2.0764-.7836-.7567-2.1215-1.0705-3.1058-.7291Zm1.5577 1.4582c.9843.3968 1.2232 1.6334.4587 2.3717-1.0704 1.0429-2.8287.1108-2.5516-1.3474.172-.8674 1.2232-1.3843 2.0929-1.0243ZM59.2973 55.5117c-.2198.406-.3823.7567-.3632.7752.0765.0646 1.2233.5998 1.2997.5998.0382 0 .2294-.3045.4205-.6829.3918-.766.4205-.6921-.4205-1.1536l-.5447-.2953-.3918.7568ZM65.0866 56.7482c-1.8157.443-3.5263 2.0119-4.2525 3.8853-.2485.6275-.4874 1.7349-.4874 2.261v.3599h1.529v-.3507c0-.1846.0669-.6275.1433-.9782.5161-2.3349 2.5707-4.0145 4.6731-3.8022.9365.0923 1.6054.4706 2.5802 1.4489.7645.7752.8314.8675.774 1.1351-.1337.6275-.0955.5999-.6976.4614-1.0703-.2491-2.2553-.1199-3.1727.3323-.1815.0922-.8696.526-1.5194.969-1.3952.9413-2.3126 1.3935-3.0294 1.4858l-.516.0553v1.4858l.5734-.0646c1.1563-.1292 2.1788-.5906 3.832-1.7442.5448-.3783 1.185-.7844 1.4239-.8951.5065-.2492 1.4717-.3415 2.1024-.1938.3154.0646.4396.0646.497-.0185.1337-.203.086.0646-.7359 4.0145-.6116 2.9347-.8123 4.079-.8123 4.6327v.7291l-.8791.8398c-.583.5445-1.1372.9782-1.6055 1.2274-1.185.646-2.6758 1.0428-3.8894 1.0428h-.5065V76.5713l.9365-.0554c2.3413-.1476 4.4915-1.0705 6.1638-2.6486l.6785-.6368.6976.6183c.9748.8583 1.4622 1.1444 2.026 1.1998.8314.0738 1.1276-.0739 2.4177-1.1998l.7072-.6183.6689.6183c1.7488 1.6612 4.2717 2.6948 6.5461 2.6948h.5638v-1.4489l-.8791-.0554c-2.0738-.1384-3.8895-.9597-5.3229-2.4179l-.6785-.6921v-.7198c0-.5261-.2102-1.7258-.8123-4.6143-.8218-3.9499-.8696-4.2175-.7358-4.0145.0573.0831.1815.0831.4969.0185.6307-.1477 1.5959-.0554 2.1024.1938.2389.1107.8792.5168 1.4239.8951 1.6532 1.1536 2.6757 1.615 3.8321 1.7442l.5733.0646v-1.4858l-.516-.0553c-.7167-.0923-1.6341-.5445-3.0294-1.4858-1.3665-.9229-1.6532-1.0798-2.4273-1.292-.6116-.1754-2.045-.1569-2.5515.0369-.2771.1015-.2867.1015-.4109-.5076-.0574-.2676.0095-.3599.774-1.1351.9748-.9783 1.6437-1.3566 2.5802-1.4489 2.112-.2123 4.157 1.4581 4.6731 3.8206.0764.3692.1433.8214.1433.9967v.3138h1.5577l-.0573-.6737c-.258-2.8332-2.2458-5.2603-4.8069-5.8417-1.2614-.2861-2.8-.0185-3.9467.6921-.2676.1661-.7167.5445-1.0034.8491l-.5256.5445H70.6579l-.4682-.4984c-1.013-1.0613-2.2171-1.615-3.6792-1.6796-.5638-.0185-1.0512.0092-1.4239.1015Zm7.9413 8.7303c.6881 3.3038 1.0703 5.3249 1.0799 5.7217v.5999l-1.013.8767c-.5638.4799-1.1181.8767-1.2423.8952-.1816.0277-.4778-.1754-1.2806-.8767l-1.0512-.9044v-.5538c0-.4614 1.9686-10.3545 2.1502-10.8251.1625-.4245.3536.2861 1.357 5.0665Z" />
+        <path d="M63.2145 67.8316c-2.1693.7567-2.8 3.3408-1.1946 4.8819.6307.6091 1.1946.8306 2.1502.8214.6881 0 .8887-.0369 1.2997-.24.7167-.3599 1.0225-.646 1.3856-1.292.2867-.526.3154-.6552.3249-1.3658 0-.9229-.2293-1.4674-.86-2.0764-.7836-.7568-2.1215-1.0706-3.1058-.7291Zm1.5576 1.4581c.9843.3968 1.2233 1.6335.4587 2.3718-.6211.5998-1.5003.5998-2.1215 0-1.1849-1.1351.1243-2.9993 1.6628-2.3718ZM78.5055 67.8316c-2.1693.7567-2.8 3.3408-1.1946 4.8819.6308.6091 1.1946.8306 2.1502.8214.6881 0 .8888-.0369 1.2997-.24.7167-.3599 1.0225-.646 1.3856-1.292.2867-.526.3154-.6552.325-1.3658 0-.9229-.2294-1.4674-.8601-2.0764-.7836-.7568-2.1215-1.0706-3.1058-.7291Zm1.5577 1.4581c.9843.3968 1.2232 1.6335.4587 2.3718-.6212.5998-1.5004.5998-2.1215 0-1.185-1.1351.1242-2.9993 1.6628-2.3718ZM66.6996 76.0368c-.2772.2584-.2963.4799-.086 1.1628.86 2.824 4.1283 4.4574 6.9379 3.47.9174-.323 1.4907-.6737 2.1406-1.3197.6211-.6183 1.0798-1.3659 1.3187-2.1503.2103-.6829.1912-.9044-.086-1.1628l-.2293-.2307h-9.7666l-.2293.2307Zm8.486 1.532c-.3823.6367-.9939 1.1904-1.7297 1.5319-.6881.3322-.7741.3507-1.6437.3507-.8696 0-.9556-.0185-1.6437-.3507-.7358-.3415-1.3474-.8952-1.7297-1.5319l-.172-.2861h7.0908l-.172.2861ZM48.8789 85.4033v.7383h1.529V84.665h-1.529v.7383ZM93.2207 85.4033v.7383h1.529V84.665h-1.529v.7383Z" />
+      </g>
+      <rect
+        width="47.9283"
+        height="46.2505"
+        x="47.8496"
+        y="40.1289"
+        stroke="red"
+        rx="23.1253"
+      />
+      <path
+        stroke="#C4C4C4"
+        d="M.7891 162.591H438.777M.7891 162.591H438.777M.7891 162.591H438.777M.7891 162.591H438.777M.7891 162.591H438.777"
+      />
+      <path
+        fill="#000"
+        d="M19.899 138.126v-2.659l6.2131-5.753c.5284-.511.9716-.972 1.3295-1.381.3637-.409.6392-.809.8267-1.201.1875-.398.2813-.827.2813-1.287 0-.512-.1165-.952-.3495-1.321-.2329-.375-.5511-.662-.9545-.861-.4034-.205-.8608-.307-1.3722-.307-.5341 0-1 .108-1.3977.324-.3977.216-.7045.526-.9204.929-.216.403-.3239.883-.3239 1.44h-3.5028c0-1.142.2585-2.133.7755-2.974.5171-.841 1.2415-1.492 2.1733-1.952.9318-.46 2.0057-.69 3.2216-.69 1.25 0 2.3381.221 3.2642.665.9318.437 1.6563 1.045 2.1733 1.823.5171.779.7756 1.671.7756 2.677 0 .659-.1307 1.309-.3921 1.951-.2557.642-.713 1.355-1.3721 2.139-.6591.779-1.5881 1.714-2.787 2.804l-2.5483 2.498v.119h7.3296v3.017H19.899ZM123.45 132.214c1.437 0 2.548.173 3.332.518.784.345 1.176.845 1.176 1.498 0 .663-.392 1.162-1.176 1.498-.784.345-1.895.518-3.332.518-1.428 0-2.534-.173-3.318-.518-.784-.336-1.176-.835-1.176-1.498 0-.653.392-1.153 1.176-1.498.784-.345 1.89-.518 3.318-.518Zm.014 1.344c-.906 0-1.568.051-1.988.154-.42.103-.63.28-.63.532 0 .233.21.406.63.518.42.103 1.082.154 1.988.154.896 0 1.549-.051 1.96-.154.42-.112.63-.285.63-.518 0-.252-.21-.429-.63-.532-.411-.103-1.064-.154-1.96-.154Zm-5.292-9.352h10.57v1.4h-10.57v-1.4Zm5.306 1.736c1.4 0 2.473.159 3.22.476.746.308 1.12.77 1.12 1.386 0 .607-.374 1.069-1.12 1.386-.747.308-1.82.462-3.22.462-1.382 0-2.45-.154-3.206-.462-.756-.317-1.134-.779-1.134-1.386 0-.616.378-1.078 1.134-1.386.756-.317 1.824-.476 3.206-.476Zm0 1.246c-.794 0-1.382.051-1.764.154-.383.093-.574.247-.574.462 0 .205.191.355.574.448.382.093.97.14 1.764.14.812 0 1.404-.047 1.778-.14.382-.093.574-.243.574-.448 0-.215-.192-.369-.574-.462-.374-.103-.966-.154-1.778-.154Zm-.924-4.018h1.848v2.002h-1.848v-2.002Zm-4.942 7.112h11.732v1.442h-11.732v-1.442Zm4.942-1.218h1.848v1.82h-1.848v-1.82Zm16.855-5.796h1.862v6.51h-1.862v-6.51Zm-4.074.49h2.044c0 1.176-.22 2.184-.658 3.024-.439.84-1.106 1.526-2.002 2.058-.887.532-2.016.938-3.388 1.218l-.644-1.498c.868-.159 1.596-.359 2.184-.602.597-.243 1.073-.527 1.428-.854.364-.327.625-.681.784-1.064.168-.383.252-.789.252-1.218v-1.064Zm-3.962 0h5.264v1.47h-5.264v-1.47Zm1.26 6.552h8.638v3.528h-6.762v1.4h-1.862v-2.744h6.748v-.756h-6.762v-1.428Zm.014 4.368h8.946v1.442h-8.946v-1.442Zm10.751-5.306h11.718v1.47h-11.718v-1.47Zm4.928-1.946h1.862v2.52h-1.862v-2.52Zm-3.514-.588h8.974v1.456h-8.974v-1.456Zm0-3.01h8.904v1.47h-7.056v2.464h-1.848v-3.934Zm4.396 7.658c1.4 0 2.501.21 3.304.63.802.411 1.204.999 1.204 1.764 0 .756-.402 1.339-1.204 1.75-.803.42-1.904.63-3.304.63s-2.502-.21-3.304-.63c-.794-.411-1.19-.994-1.19-1.75 0-.765.396-1.353 1.19-1.764.802-.42 1.904-.63 3.304-.63Zm0 1.414c-.579 0-1.064.037-1.456.112-.383.065-.672.173-.868.322-.196.14-.294.317-.294.532 0 .224.098.411.294.56.196.14.485.247.868.322.392.065.877.098 1.456.098.578 0 1.059-.033 1.442-.098.392-.075.686-.182.882-.322.196-.149.294-.336.294-.56 0-.215-.098-.392-.294-.532-.196-.149-.49-.257-.882-.322-.383-.075-.864-.112-1.442-.112ZM226.562 123.818V134h-2.153v-8.138h-.06l-2.331 1.461v-1.909l2.52-1.596h2.024Zm6.412 10.406c-.855-.004-1.591-.214-2.208-.632-.613-.417-1.085-1.022-1.417-1.814-.328-.792-.49-1.745-.487-2.859 0-1.11.164-2.057.492-2.839.332-.782.804-1.377 1.417-1.785.617-.411 1.351-.616 2.203-.616.852 0 1.584.205 2.197.616.617.411 1.091 1.008 1.422 1.79.331.779.496 1.724.492 2.834 0 1.117-.165 2.072-.497 2.864-.328.792-.799 1.397-1.412 1.814-.613.418-1.347.627-2.202.627Zm0-1.785c.583 0 1.049-.293 1.397-.88.348-.587.52-1.467.517-2.64 0-.772-.08-1.415-.239-1.929-.156-.514-.378-.9-.666-1.158-.285-.259-.621-.388-1.009-.388-.58 0-1.044.29-1.392.87-.348.58-.524 1.448-.527 2.605 0 .782.078 1.435.233 1.959.159.52.383.911.672 1.173.288.259.626.388 1.014.388Zm9.638 1.785c-.855-.004-1.591-.214-2.207-.632-.613-.417-1.085-1.022-1.417-1.814-.328-.792-.49-1.745-.487-2.859 0-1.11.164-2.057.492-2.839.332-.782.804-1.377 1.417-1.785.616-.411 1.351-.616 2.202-.616.852 0 1.585.205 2.198.616.616.411 1.09 1.008 1.422 1.79.331.779.495 1.724.492 2.834 0 1.117-.166 2.072-.497 2.864-.328.792-.799 1.397-1.412 1.814-.613.418-1.347.627-2.203.627Zm0-1.785c.584 0 1.049-.293 1.397-.88.348-.587.521-1.467.517-2.64 0-.772-.079-1.415-.238-1.929-.156-.514-.378-.9-.666-1.158-.285-.259-.622-.388-1.01-.388-.58 0-1.044.29-1.392.87-.348.58-.523 1.448-.527 2.605 0 .782.078 1.435.234 1.959.159.52.383.911.671 1.173.289.259.627.388 1.014.388ZM306.053 124.516v10.182H303.9v-8.138h-.059l-2.332 1.461v-1.909l2.521-1.596h2.023ZM354.322 132v-1.551l3.625-3.356c.308-.298.566-.567.775-.805.212-.239.373-.473.482-.701.11-.232.165-.483.165-.751 0-.298-.068-.555-.204-.771-.136-.218-.322-.386-.557-.502-.235-.119-.502-.179-.801-.179-.311 0-.583.063-.815.189-.232.126-.411.307-.537.542-.126.235-.189.515-.189.84h-2.043c0-.666.151-1.244.452-1.735.302-.49.725-.87 1.268-1.138.544-.269 1.17-.403 1.879-.403.73 0 1.364.129 1.904.388.544.255.967.61 1.268 1.064.302.454.453.974.453 1.561 0 .384-.077.764-.229 1.138-.149.375-.416.791-.8 1.248-.385.454-.927.999-1.626 1.636l-1.487 1.456v.07h4.276V132h-7.259Zm12.811.224c-.855-.004-1.591-.214-2.207-.632-.614-.417-1.086-1.022-1.417-1.814-.329-.792-.491-1.745-.488-2.859 0-1.11.164-2.057.493-2.839.331-.782.803-1.377 1.416-1.785.617-.411 1.351-.616 2.203-.616.852 0 1.584.205 2.197.616.617.411 1.091 1.008 1.422 1.79.332.779.496 1.724.492 2.834 0 1.117-.165 2.072-.497 2.864-.328.792-.799 1.397-1.412 1.814-.613.418-1.347.627-2.202.627Zm0-1.785c.583 0 1.049-.293 1.397-.88.348-.587.52-1.467.517-2.64 0-.772-.08-1.415-.239-1.929-.155-.514-.378-.9-.666-1.158-.285-.259-.621-.388-1.009-.388-.58 0-1.044.29-1.392.87-.348.58-.524 1.448-.527 2.605 0 .782.078 1.435.234 1.959.159.52.382.911.671 1.173.288.259.626.388 1.014.388Zm5.646 1.561v-1.551l3.625-3.356c.308-.298.566-.567.775-.805.212-.239.373-.473.482-.701.11-.232.165-.483.165-.751 0-.298-.068-.555-.204-.771-.136-.218-.322-.386-.557-.502-.235-.119-.502-.179-.801-.179-.311 0-.583.063-.815.189-.232.126-.411.307-.537.542-.126.235-.189.515-.189.84h-2.043c0-.666.151-1.244.452-1.735.302-.49.725-.87 1.268-1.138.544-.269 1.17-.403 1.879-.403.73 0 1.364.129 1.905.388.543.255.966.61 1.267 1.064.302.454.453.974.453 1.561 0 .384-.077.764-.229 1.138-.149.375-.416.791-.8 1.248-.385.454-.927.999-1.626 1.636l-1.487 1.456v.07h4.276V132h-7.259Zm12.537.139c-.702 0-1.329-.129-1.879-.388-.547-.258-.981-.614-1.302-1.068-.322-.455-.489-.975-.502-1.562h2.088c.023.395.189.715.497.96s.674.368 1.098.368c.339 0 .637-.075.895-.224.262-.152.466-.363.612-.631.149-.272.224-.584.224-.935 0-.358-.077-.673-.229-.945-.149-.271-.356-.483-.622-.636-.265-.152-.568-.23-.909-.234-.299 0-.589.062-.87.184-.279.123-.496.29-.652.503l-1.914-.343.483-5.37h6.224v1.76h-4.45l-.263 2.551h.06c.179-.252.449-.461.81-.627.361-.166.766-.248 1.213-.248.613 0 1.16.144 1.641.432.48.288.86.685 1.138 1.188.279.501.416 1.077.413 1.73.003.686-.156 1.296-.478 1.83-.318.53-.763.948-1.337 1.253-.57.301-1.233.452-1.989.452Zm6.235-.01c-.328 0-.61-.116-.845-.348-.232-.235-.348-.517-.348-.845 0-.325.116-.603.348-.835.235-.232.517-.348.845-.348.318 0 .596.116.835.348.239.232.358.51.358.835 0 .219-.056.419-.169.602-.109.179-.254.323-.433.432-.178.106-.376.159-.591.159Zm6.494.095c-.855-.004-1.591-.214-2.207-.632-.614-.417-1.086-1.022-1.417-1.814-.328-.792-.491-1.745-.487-2.859 0-1.11.164-2.057.492-2.839.331-.782.803-1.377 1.417-1.785.616-.411 1.35-.616 2.202-.616.852 0 1.584.205 2.197.616.617.411 1.091 1.008 1.422 1.79.332.779.496 1.724.493 2.834 0 1.117-.166 2.072-.498 2.864-.328.792-.798 1.397-1.412 1.814-.613.418-1.347.627-2.202.627Zm0-1.785c.583 0 1.049-.293 1.397-.88.348-.587.52-1.467.517-2.64 0-.772-.079-1.415-.239-1.929-.155-.514-.377-.9-.666-1.158-.285-.259-.621-.388-1.009-.388-.58 0-1.044.29-1.392.87-.348.58-.524 1.448-.527 2.605 0 .782.078 1.435.234 1.959.159.52.382.911.671 1.173.288.259.626.388 1.014.388Zm10.076-8.621V132h-2.153v-8.138h-.059l-2.332 1.461v-1.909l2.521-1.596h2.023Zm3.131 10.311c-.328 0-.61-.116-.845-.348-.232-.235-.348-.517-.348-.845 0-.325.116-.603.348-.835.235-.232.517-.348.845-.348.318 0 .597.116.835.348.239.232.358.51.358.835 0 .219-.056.419-.169.602-.109.179-.253.323-.432.432-.179.106-.377.159-.592.159Zm6.07-10.311V132h-2.152v-8.138h-.06l-2.332 1.461v-1.909l2.521-1.596h2.023ZM420.145 132l4.221-8.352v-.07h-4.917v-1.76h7.144v1.785L422.367 132h-2.222Z"
+      />
+      <g fill="red" clipPath="url(#b)">
+        <path d="M61.025 105.318c-1.1277.286-2.1884 1.089-2.6949 2.021-.2485.443-3.3638 8.887-3.6601 9.921-.3631 1.227-.4396 2.261-.4396 5.832v3.489h-.3631c-.8792 0-1.9399.572-2.3031 1.236-.1051.176-.2102.24-.4204.24-.8506 0-1.9113.757-2.1502 1.532-.086.286-.1147 2.667-.1147 9.764v9.376h1.529V129.986l.2389-.231c.3441-.332.5256-.258.6881.277.2867.932 1.4334 1.717 2.5324 1.717h.3631v2.104c0 2.335.0765 2.787.5925 3.728.4014.729 1.3093 1.541 2.112 1.874.6307.267.6403.276.7932.784.086.277.344.987.5829 1.569 1.099 2.704 2.7235 5.113 4.6348 6.884 7.0621 6.544 16.7427 4.218 21.6068-5.186.2389-.452.4492-.904.4683-.988.0191-.129-.1338-.23-.6307-.433-.3632-.148-.669-.268-.6881-.249-.0095.009-.2962.553-.6402 1.218-2.284 4.356-5.6861 7.078-9.6615 7.752-2.4655.415-5.2273-.194-7.4635-1.643-3.985-2.584-6.7658-7.411-7.6546-13.308-.2389-1.559-.2389-5.205-.0095-6.69.2389-1.468.602-3.009.9843-4.162.1815-.536.3249-1.016.3249-1.062 0-.046-.3058-.194-.6785-.332l-.6785-.24-.1338.323c-.4396 1.006-1.1563 3.931-1.3761 5.565-.2867 2.187-.2103 5.353.1816 7.567.0955.536.1624.988.1433 1.006-.0764.083-.7836-.729-1.0034-1.153l-.2294-.452v-5.63c0-5.989.0191-6.303.4874-8.112.8505-3.313 2.6758-6.358 5.1413-8.61 2.3222-2.123 4.7686-3.396 7.7502-4.015 1.3761-.286 3.8894-.286 5.2655 0 1.7297.36 3.5167 1.043 4.7973 1.828.2771.175.516.313.5351.313.0191 0 .2198-.276.4492-.609.3345-.498.3822-.618.2676-.692a40.8223 40.8223 0 0 0-.9843-.563l-.841-.48.0478-.516c.086-.933.6594-1.745 1.5481-2.197.6594-.341 1.7966-.351 2.4369-.028.5542.277 1.0225.702 1.2518 1.145.1052.194.927 2.353 1.8157 4.799 1.4144 3.866 1.9878 5.629 1.9973 6.155 0 .074-.2484-.341-.5543-.923-.6784-1.301-1.3761-2.344-2.3795-3.553-.8123-.969-2.4655-2.574-2.5993-2.519-.0478.018-.2771.258-.516.535l-.4396.508 1.1276 1.107c.6116.609 1.3188 1.357 1.5577 1.661 2.1406 2.741 3.4498 5.925 3.8034 9.321.0669.665.0956 2.778.0765 6.045l-.0287 5.03-.3249.581c-.2772.489-.8314 1.107-.9174 1.015-.0096-.018.0573-.471.1529-1.006 1.3187-7.457-.8983-15.375-5.6287-20.137-3.6696-3.673-8.4669-4.983-12.9679-3.534-1.9495.637-3.4881 1.559-5.1509 3.101-1.1276 1.052-2.3317 2.464-2.1979 2.574.2771.231 1.0798.748 1.1467.739.0478 0 .4301-.425.8601-.942 2.7809-3.396 6.7468-5.057 10.598-4.448 3.2969.517 6.2498 2.538 8.5146 5.814 3.4594 5.03 4.568 12.2 2.8669 18.642-.1338.508-.3058 1.098-.3822 1.31-.086.231-.1051.416-.0574.443.1529.093 1.1659.416 1.2901.416.0669 0 .1912-.24.2772-.536.1529-.526.1529-.526.7836-.793 1.2805-.545 2.3604-1.855 2.6184-3.184.0574-.268.0956-1.403.0956-2.51v-2.021h.3631c1.099 0 2.2458-.785 2.5325-1.717.1624-.535.344-.609.688-.277l.2389.231V148.729h1.529v-9.376c0-6.986-.0286-9.478-.1146-9.745-.2485-.822-1.2615-1.551-2.1311-1.551-.2294 0-.3345-.055-.4396-.24-.3631-.664-1.4239-1.236-2.3031-1.236h-.3631v-3.489c0-3.571-.0765-4.605-.4396-5.832-.2962-1.034-3.4116-9.478-3.6601-9.921-.516-.951-1.5768-1.735-2.7331-2.03-2.3508-.591-4.8068.904-5.3037 3.239-.0956.415-.1434.498-.2676.452-.4683-.175-1.8539-.526-2.628-.664-1.2041-.213-3.899-.222-5.084-.01-.7549.139-2.1788.499-2.6471.674-.1242.046-.172-.037-.2675-.452-.5065-2.344-2.9816-3.848-5.342-3.23Zm2.2839 1.606c.9557.461 1.5482 1.319 1.5673 2.242l.0095.545-.8983.498c-1.3761.766-2.3317 1.467-3.5645 2.612-1.7488 1.624-3.0006 3.248-4.0614 5.26-.2771.526-.5065.923-.5065.877 0-.046.086-.434.1816-.868.1911-.812 3.2969-9.413 3.6314-10.031.3058-.582.9652-1.089 1.7297-1.329.4205-.139 1.4525-.037 1.9112.194Zm-9.0785 22.25v1.126l-.4395-.056c-.6594-.064-1.0417-.47-1.0417-1.07 0-.674.4492-1.098 1.1754-1.108l.3058-.009v1.117Zm36.3523-.757c.2484.249.2962.36.2962.748 0 .609-.3822 1.015-1.0416 1.079l-.4396.056v-2.27l.4396.055c.3249.028.5256.12.7454.332Z" />
+        <path d="M70.86 115.654c-2.1693.757-2.8 3.341-1.1946 4.882.6308.609 1.1946.83 2.1502.821.6881 0 .8887-.037 1.2997-.24.7167-.36 1.0225-.646 1.3856-1.292.2867-.526.3154-.655.3249-1.366 0-.923-.2293-1.467-.86-2.076-.7836-.757-2.1215-1.071-3.1058-.729Zm1.5577 1.458c.9843.397 1.2232 1.633.4587 2.372-1.0704 1.043-2.8287.11-2.5516-1.348.172-.867 1.2232-1.384 2.0929-1.024ZM59.2973 121.053c-.2198.406-.3823.756-.3632.775.0765.064 1.2233.6 1.2997.6.0382 0 .2294-.305.4205-.683.3918-.766.4205-.692-.4205-1.154l-.5447-.295-.3918.757ZM65.0866 122.289c-1.8157.443-3.5263 2.012-4.2525 3.885-.2485.628-.4874 1.735-.4874 2.261v.36h1.529v-.35c0-.185.0669-.628.1433-.979.5161-2.334 2.5707-4.014 4.6731-3.802.9365.093 1.6054.471 2.5802 1.449.7645.775.8314.868.774 1.135-.1337.628-.0955.6-.6976.462-1.0703-.249-2.2553-.12-3.1727.332-.1815.092-.8696.526-1.5194.969-1.3952.941-2.3126 1.393-3.0294 1.486l-.516.055v1.486l.5734-.065c1.1563-.129 2.1788-.59 3.832-1.744.5448-.378 1.185-.784 1.4239-.895.5065-.249 1.4717-.341 2.1024-.194.3154.065.4396.065.497-.018.1337-.203.086.064-.7359 4.014-.6116 2.935-.8123 4.079-.8123 4.633v.729l-.8791.84c-.583.544-1.1372.978-1.6055 1.227-1.185.646-2.6758 1.043-3.8894 1.043h-.5065v1.504l.9365-.055c2.3413-.148 4.4915-1.071 6.1638-2.649l.6785-.636.6976.618c.9748.858 1.4622 1.144 2.026 1.2.8314.073 1.1276-.074 2.4177-1.2l.7072-.618.6689.618c1.7488 1.661 4.2717 2.695 6.5461 2.695h.5638V140.636l-.8791-.056c-2.0738-.138-3.8895-.959-5.3229-2.418l-.6785-.692v-.72c0-.526-.2102-1.725-.8123-4.614-.8218-3.95-.8696-4.217-.7358-4.014.0573.083.1815.083.4969.018.6307-.147 1.5959-.055 2.1024.194.2389.111.8792.517 1.4239.895 1.6532 1.154 2.6757 1.615 3.8321 1.744l.5733.065v-1.486l-.516-.055c-.7167-.093-1.6341-.545-3.0294-1.486-1.3665-.923-1.6532-1.08-2.4273-1.292-.6116-.175-2.045-.157-2.5515.037-.2771.101-.2867.101-.4109-.508-.0574-.267.0095-.36.774-1.135.9748-.978 1.6437-1.356 2.5802-1.449 2.112-.212 4.157 1.458 4.6731 3.821.0764.369.1433.821.1433.997v.313h1.5577l-.0573-.673c-.258-2.833-2.2458-5.261-4.8069-5.842-1.2614-.286-2.8-.018-3.9467.692-.2676.166-.7167.545-1.0034.849l-.5256.545H70.6579l-.4682-.499c-1.013-1.061-2.2171-1.615-3.6792-1.679-.5638-.019-1.0512.009-1.4239.101Zm7.9413 8.731c.6881 3.303 1.0703 5.324 1.0799 5.721v.6l-1.013.877c-.5638.48-1.1181.877-1.2423.895-.1816.028-.4778-.175-1.2806-.877l-1.0512-.904v-.554c0-.461 1.9686-10.354 2.1502-10.825.1625-.425.3536.286 1.357 5.067Z" />
+        <path d="M63.2145 133.373c-2.1693.756-2.8 3.34-1.1946 4.882.6307.609 1.1946.83 2.1502.821.6881 0 .8887-.037 1.2997-.24.7167-.36 1.0225-.646 1.3856-1.292.2867-.526.3154-.655.3249-1.366 0-.923-.2293-1.467-.86-2.076-.7836-.757-2.1215-1.071-3.1058-.729Zm1.5576 1.458c.9843.397 1.2233 1.633.4587 2.371-.6211.6-1.5003.6-2.1215 0-1.1849-1.135.1243-2.999 1.6628-2.371ZM78.5055 133.373c-2.1693.756-2.8 3.34-1.1946 4.882.6308.609 1.1946.83 2.1502.821.6881 0 .8888-.037 1.2997-.24.7167-.36 1.0225-.646 1.3856-1.292.2867-.526.3154-.655.325-1.366 0-.923-.2294-1.467-.8601-2.076-.7836-.757-2.1215-1.071-3.1058-.729Zm1.5577 1.458c.9843.397 1.2232 1.633.4587 2.371-.6212.6-1.5004.6-2.1215 0-1.185-1.135.1242-2.999 1.6628-2.371ZM66.6996 141.578c-.2772.258-.2963.48-.086 1.163.86 2.824 4.1283 4.457 6.9379 3.47.9174-.323 1.4907-.674 2.1406-1.32.6211-.618 1.0798-1.366 1.3187-2.15.2103-.683.1912-.905-.086-1.163l-.2293-.231h-9.7666l-.2293.231Zm8.486 1.532c-.3823.637-.9939 1.19-1.7297 1.532-.6881.332-.7741.35-1.6437.35-.8696 0-.9556-.018-1.6437-.35-.7358-.342-1.3474-.895-1.7297-1.532l-.172-.286h7.0908l-.172.286ZM48.8789 150.944v.739h1.529V150.206h-1.529v.738ZM93.2207 150.944v.739h1.529V150.206h-1.529v.738Z" />
+      </g>
+      <rect
+        width="47.9283"
+        height="46.2505"
+        x="47.8496"
+        y="105.67"
+        stroke="red"
+        rx="23.1253"
+      />
+      <path
+        stroke="#C4C4C4"
+        d="M.7891 228.133H438.777M.7891 228.133H438.777M.7891 228.133H438.777M.7891 228.133H438.777M.7891 228.133H438.777"
+      />
+      <path
+        fill="#000"
+        d="M25.9991 203.907c-1.2727 0-2.4062-.219-3.4005-.657-.9887-.443-1.7699-1.051-2.3438-1.824-.5682-.778-.8608-1.676-.8778-2.693h3.7159c.0227.426.1619.801.4176 1.125.2614.318.608.566 1.0398.742.4318.176.9176.264 1.4574.264.5625 0 1.0596-.099 1.4914-.298.4319-.199.7699-.475 1.0143-.827.2443-.352.3664-.759.3664-1.219 0-.466-.1307-.878-.392-1.236-.2557-.363-.625-.647-1.108-.852-.4772-.204-1.0454-.307-1.7045-.307h-1.6279v-2.71h1.6279c.5568 0 1.0483-.096 1.4744-.29.4318-.193.7671-.46 1.0057-.801.2386-.346.358-.75.358-1.21 0-.438-.1052-.821-.3154-1.151-.2045-.335-.4943-.596-.8693-.784-.3693-.187-.8011-.281-1.2955-.281-.5 0-.9573.091-1.3721.273-.4148.176-.7472.429-.9972.758-.25.33-.3835.716-.4005 1.159h-3.537c.0171-1.005.304-1.892.8608-2.659.5568-.767 1.3068-1.366 2.25-1.798.9489-.437 2.0199-.656 3.2131-.656 1.2045 0 2.2585.219 3.1619.656.9034.438 1.6051 1.028 2.1051 1.773.5057.738.7557 1.568.75 2.488.0057.978-.2983 1.793-.9119 2.446-.608.654-1.4006 1.069-2.3778 1.245v.136c1.284.165 2.2613.611 2.9318 1.338.6761.722 1.0113 1.625 1.0057 2.71.0056.995-.2813 1.878-.8608 2.651-.5739.773-1.3665 1.381-2.3779 1.824-1.0113.443-2.1704.665-3.4773.665ZM123.45 197.214c1.437 0 2.548.173 3.332.518.784.345 1.176.845 1.176 1.498 0 .663-.392 1.162-1.176 1.498-.784.345-1.895.518-3.332.518-1.428 0-2.534-.173-3.318-.518-.784-.336-1.176-.835-1.176-1.498 0-.653.392-1.153 1.176-1.498.784-.345 1.89-.518 3.318-.518Zm.014 1.344c-.906 0-1.568.051-1.988.154-.42.103-.63.28-.63.532 0 .233.21.406.63.518.42.103 1.082.154 1.988.154.896 0 1.549-.051 1.96-.154.42-.112.63-.285.63-.518 0-.252-.21-.429-.63-.532-.411-.103-1.064-.154-1.96-.154Zm-5.292-9.352h10.57v1.4h-10.57v-1.4Zm5.306 1.736c1.4 0 2.473.159 3.22.476.746.308 1.12.77 1.12 1.386 0 .607-.374 1.069-1.12 1.386-.747.308-1.82.462-3.22.462-1.382 0-2.45-.154-3.206-.462-.756-.317-1.134-.779-1.134-1.386 0-.616.378-1.078 1.134-1.386.756-.317 1.824-.476 3.206-.476Zm0 1.246c-.794 0-1.382.051-1.764.154-.383.093-.574.247-.574.462 0 .205.191.355.574.448.382.093.97.14 1.764.14.812 0 1.404-.047 1.778-.14.382-.093.574-.243.574-.448 0-.215-.192-.369-.574-.462-.374-.103-.966-.154-1.778-.154Zm-.924-4.018h1.848v2.002h-1.848v-2.002Zm-4.942 7.112h11.732v1.442h-11.732v-1.442Zm4.942-1.218h1.848v1.82h-1.848v-1.82Zm16.855-5.796h1.862v6.51h-1.862v-6.51Zm-4.074.49h2.044c0 1.176-.22 2.184-.658 3.024-.439.84-1.106 1.526-2.002 2.058-.887.532-2.016.938-3.388 1.218l-.644-1.498c.868-.159 1.596-.359 2.184-.602.597-.243 1.073-.527 1.428-.854.364-.327.625-.681.784-1.064.168-.383.252-.789.252-1.218v-1.064Zm-3.962 0h5.264v1.47h-5.264v-1.47Zm1.26 6.552h8.638v3.528h-6.762v1.4h-1.862v-2.744h6.748v-.756h-6.762v-1.428Zm.014 4.368h8.946v1.442h-8.946v-1.442Zm10.751-5.306h11.718v1.47h-11.718v-1.47Zm4.928-1.946h1.862v2.52h-1.862v-2.52Zm-3.514-.588h8.974v1.456h-8.974v-1.456Zm0-3.01h8.904v1.47h-7.056v2.464h-1.848v-3.934Zm4.396 7.658c1.4 0 2.501.21 3.304.63.802.411 1.204.999 1.204 1.764 0 .756-.402 1.339-1.204 1.75-.803.42-1.904.63-3.304.63s-2.502-.21-3.304-.63c-.794-.411-1.19-.994-1.19-1.75 0-.765.396-1.353 1.19-1.764.802-.42 1.904-.63 3.304-.63Zm0 1.414c-.579 0-1.064.037-1.456.112-.383.065-.672.173-.868.322-.196.14-.294.317-.294.532 0 .224.098.411.294.56.196.14.485.247.868.322.392.065.877.098 1.456.098.578 0 1.059-.033 1.442-.098.392-.075.686-.182.882-.322.196-.149.294-.336.294-.56 0-.215-.098-.392-.294-.532-.196-.149-.49-.257-.882-.322-.383-.075-.864-.112-1.442-.112ZM226.562 188.818V199h-2.153v-8.138h-.06l-2.331 1.461v-1.909l2.52-1.596h2.024Zm6.412 10.406c-.855-.004-1.591-.214-2.208-.632-.613-.417-1.085-1.022-1.417-1.814-.328-.792-.49-1.745-.487-2.859 0-1.11.164-2.057.492-2.839.332-.782.804-1.377 1.417-1.785.617-.411 1.351-.616 2.203-.616.852 0 1.584.205 2.197.616.617.411 1.091 1.008 1.422 1.79.331.779.496 1.724.492 2.834 0 1.117-.165 2.072-.497 2.864-.328.792-.799 1.397-1.412 1.814-.613.418-1.347.627-2.202.627Zm0-1.785c.583 0 1.049-.293 1.397-.88.348-.587.52-1.467.517-2.64 0-.772-.08-1.415-.239-1.929-.156-.514-.378-.9-.666-1.158-.285-.259-.621-.388-1.009-.388-.58 0-1.044.29-1.392.87-.348.58-.524 1.448-.527 2.605 0 .782.078 1.435.233 1.959.159.52.383.911.672 1.173.288.259.626.388 1.014.388Zm9.638 1.785c-.855-.004-1.591-.214-2.207-.632-.613-.417-1.085-1.022-1.417-1.814-.328-.792-.49-1.745-.487-2.859 0-1.11.164-2.057.492-2.839.332-.782.804-1.377 1.417-1.785.616-.411 1.351-.616 2.202-.616.852 0 1.585.205 2.198.616.616.411 1.09 1.008 1.422 1.79.331.779.495 1.724.492 2.834 0 1.117-.166 2.072-.497 2.864-.328.792-.799 1.397-1.412 1.814-.613.418-1.347.627-2.203.627Zm0-1.785c.584 0 1.049-.293 1.397-.88.348-.587.521-1.467.517-2.64 0-.772-.079-1.415-.238-1.929-.156-.514-.378-.9-.666-1.158-.285-.259-.622-.388-1.01-.388-.58 0-1.044.29-1.392.87-.348.58-.523 1.448-.527 2.605 0 .782.078 1.435.234 1.959.159.52.383.911.671 1.173.289.259.627.388 1.014.388ZM306.053 190.058v10.182H303.9v-8.138h-.059l-2.332 1.461v-1.909l2.521-1.596h2.023ZM353.822 198v-1.551l3.625-3.356c.308-.298.566-.567.775-.805.212-.239.373-.473.482-.701.11-.232.165-.483.165-.751 0-.298-.068-.555-.204-.771-.136-.218-.322-.386-.557-.502-.235-.119-.502-.179-.801-.179-.311 0-.583.063-.815.189-.232.126-.411.307-.537.542-.126.235-.189.515-.189.84h-2.043c0-.666.151-1.244.452-1.735.302-.49.725-.87 1.268-1.138.544-.269 1.17-.403 1.879-.403.73 0 1.364.129 1.904.388.544.255.967.61 1.268 1.064.302.454.453.974.453 1.561 0 .384-.077.764-.229 1.138-.149.375-.416.791-.8 1.248-.385.454-.927.999-1.626 1.636l-1.487 1.456v.07h4.276V198h-7.259Zm12.811.224c-.855-.004-1.591-.214-2.207-.632-.614-.417-1.086-1.022-1.417-1.814-.329-.792-.491-1.745-.488-2.859 0-1.11.164-2.057.493-2.839.331-.782.803-1.377 1.416-1.785.617-.411 1.351-.616 2.203-.616.852 0 1.584.205 2.197.616.617.411 1.091 1.008 1.422 1.79.332.779.496 1.724.492 2.834 0 1.117-.165 2.072-.497 2.864-.328.792-.799 1.397-1.412 1.814-.613.418-1.347.627-2.202.627Zm0-1.785c.583 0 1.049-.293 1.397-.88.348-.587.52-1.467.517-2.64 0-.772-.08-1.415-.239-1.929-.155-.514-.378-.9-.666-1.158-.285-.259-.621-.388-1.009-.388-.58 0-1.044.29-1.392.87-.348.58-.524 1.448-.527 2.605 0 .782.078 1.435.234 1.959.159.52.382.911.671 1.173.288.259.626.388 1.014.388Zm5.646 1.561v-1.551l3.625-3.356c.308-.298.566-.567.775-.805.212-.239.373-.473.482-.701.11-.232.165-.483.165-.751 0-.298-.068-.555-.204-.771-.136-.218-.322-.386-.557-.502-.235-.119-.502-.179-.801-.179-.311 0-.583.063-.815.189-.232.126-.411.307-.537.542-.126.235-.189.515-.189.84h-2.043c0-.666.151-1.244.452-1.735.302-.49.725-.87 1.268-1.138.544-.269 1.17-.403 1.879-.403.73 0 1.364.129 1.905.388.543.255.966.61 1.267 1.064.302.454.453.974.453 1.561 0 .384-.077.764-.229 1.138-.149.375-.416.791-.8 1.248-.385.454-.927.999-1.626 1.636l-1.487 1.456v.07h4.276V198h-7.259Zm12.537.139c-.702 0-1.329-.129-1.879-.388-.547-.258-.981-.614-1.302-1.068-.322-.455-.489-.975-.502-1.562h2.088c.023.395.189.715.497.96s.674.368 1.098.368c.339 0 .637-.075.895-.224.262-.152.466-.363.612-.631.149-.272.224-.584.224-.935 0-.358-.077-.673-.229-.945-.149-.271-.356-.483-.622-.636-.265-.152-.568-.23-.909-.234-.299 0-.589.062-.87.184-.279.123-.496.29-.652.503l-1.914-.343.483-5.37h6.224v1.76h-4.45l-.263 2.551h.06c.179-.252.449-.461.81-.627.361-.166.766-.248 1.213-.248.613 0 1.16.144 1.641.432.48.288.86.685 1.138 1.188.279.501.416 1.077.413 1.73.003.686-.156 1.296-.478 1.83-.318.53-.763.948-1.337 1.253-.57.301-1.233.452-1.989.452Zm6.235-.01c-.328 0-.61-.116-.845-.348-.232-.235-.348-.517-.348-.845 0-.325.116-.603.348-.835.235-.232.517-.348.845-.348.318 0 .596.116.835.348.239.232.358.51.358.835 0 .219-.056.419-.169.602-.109.179-.254.323-.433.432-.178.106-.376.159-.591.159Zm6.494.095c-.855-.004-1.591-.214-2.207-.632-.614-.417-1.086-1.022-1.417-1.814-.328-.792-.491-1.745-.487-2.859 0-1.11.164-2.057.492-2.839.331-.782.803-1.377 1.417-1.785.616-.411 1.35-.616 2.202-.616.852 0 1.584.205 2.197.616.617.411 1.091 1.008 1.422 1.79.332.779.496 1.724.493 2.834 0 1.117-.166 2.072-.498 2.864-.328.792-.798 1.397-1.412 1.814-.613.418-1.347.627-2.202.627Zm0-1.785c.583 0 1.049-.293 1.397-.88.348-.587.52-1.467.517-2.64 0-.772-.079-1.415-.239-1.929-.155-.514-.377-.9-.666-1.158-.285-.259-.621-.388-1.009-.388-.58 0-1.044.29-1.392.87-.348.58-.524 1.448-.527 2.605 0 .782.078 1.435.234 1.959.159.52.382.911.671 1.173.288.259.626.388 1.014.388Zm10.076-8.621V198h-2.153v-8.138h-.059l-2.332 1.461v-1.909l2.521-1.596h2.023Zm3.131 10.311c-.328 0-.61-.116-.845-.348-.232-.235-.348-.517-.348-.845 0-.325.116-.603.348-.835.235-.232.517-.348.845-.348.318 0 .597.116.835.348.239.232.358.51.358.835 0 .219-.056.419-.169.602-.109.179-.253.323-.432.432-.179.106-.377.159-.592.159Zm6.07-10.311V198h-2.152v-8.138h-.06l-2.332 1.461v-1.909l2.521-1.596h2.023ZM419.645 198l4.221-8.352v-.07h-4.917v-1.76h7.144v1.785L421.867 198h-2.222Z"
+      />
+      <g fill="red" clipPath="url(#c)">
+        <path d="M61.025 170.86c-1.1277.286-2.1884 1.089-2.6949 2.021-.2485.443-3.3638 8.887-3.6601 9.921-.3631 1.227-.4396 2.261-.4396 5.832v3.489h-.3631c-.8792 0-1.9399.572-2.3031 1.236-.1051.176-.2102.24-.4204.24-.8506 0-1.9113.757-2.1502 1.532-.086.286-.1147 2.667-.1147 9.764v9.376h1.529V195.528l.2389-.231c.3441-.332.5256-.258.6881.277.2867.932 1.4334 1.717 2.5324 1.717h.3631v2.104c0 2.335.0765 2.787.5925 3.728.4014.729 1.3093 1.541 2.112 1.874.6307.267.6403.276.7932.784.086.277.344.987.5829 1.569 1.099 2.704 2.7235 5.113 4.6348 6.884 7.0621 6.543 16.7427 4.218 21.6068-5.186.2389-.452.4492-.904.4683-.988.0191-.129-.1338-.23-.6307-.433-.3632-.148-.669-.268-.6881-.249-.0095.009-.2962.553-.6402 1.218-2.284 4.356-5.6861 7.078-9.6615 7.752-2.4655.415-5.2273-.194-7.4635-1.643-3.985-2.584-6.7658-7.411-7.6546-13.308-.2389-1.559-.2389-5.205-.0095-6.69.2389-1.468.602-3.009.9843-4.162.1815-.536.3249-1.016.3249-1.062 0-.046-.3058-.194-.6785-.332l-.6785-.24-.1338.323c-.4396 1.006-1.1563 3.931-1.3761 5.565-.2867 2.187-.2103 5.353.1816 7.567.0955.536.1624.988.1433 1.006-.0764.083-.7836-.729-1.0034-1.153l-.2294-.452v-5.63c0-5.989.0191-6.303.4874-8.112.8505-3.313 2.6758-6.358 5.1413-8.61 2.3222-2.123 4.7686-3.396 7.7502-4.015 1.3761-.286 3.8894-.286 5.2655 0 1.7297.36 3.5167 1.043 4.7973 1.828.2771.175.516.313.5351.313.0191 0 .2198-.276.4492-.609.3345-.498.3822-.618.2676-.692a40.8223 40.8223 0 0 0-.9843-.563l-.841-.48.0478-.516c.086-.933.6594-1.745 1.5481-2.197.6594-.341 1.7966-.351 2.4369-.028.5542.277 1.0225.702 1.2518 1.145.1052.194.927 2.353 1.8157 4.799 1.4144 3.866 1.9878 5.629 1.9973 6.155 0 .074-.2484-.341-.5543-.923-.6784-1.301-1.3761-2.344-2.3795-3.553-.8123-.969-2.4655-2.574-2.5993-2.519-.0478.018-.2771.258-.516.535l-.4396.508 1.1276 1.107c.6116.609 1.3188 1.357 1.5577 1.661 2.1406 2.741 3.4498 5.925 3.8034 9.321.0669.665.0956 2.778.0765 6.045l-.0287 5.03-.3249.581c-.2772.489-.8314 1.107-.9174 1.015-.0096-.018.0573-.471.1529-1.006 1.3187-7.457-.8983-15.375-5.6287-20.137-3.6696-3.673-8.4669-4.983-12.9679-3.534-1.9495.637-3.4881 1.559-5.1509 3.101-1.1276 1.052-2.3317 2.464-2.1979 2.574.2771.231 1.0798.748 1.1467.739.0478 0 .4301-.425.8601-.942 2.7809-3.396 6.7468-5.057 10.598-4.448 3.2969.517 6.2498 2.538 8.5146 5.814 3.4594 5.03 4.568 12.2 2.8669 18.642-.1338.508-.3058 1.098-.3822 1.31-.086.231-.1051.416-.0574.443.1529.093 1.1659.416 1.2901.416.0669 0 .1912-.24.2772-.536.1529-.526.1529-.526.7836-.793 1.2805-.545 2.3604-1.855 2.6184-3.184.0574-.268.0956-1.403.0956-2.51v-2.021h.3631c1.099 0 2.2458-.785 2.5325-1.717.1624-.535.344-.609.688-.277l.2389.231V214.271h1.529v-9.376c0-6.986-.0286-9.478-.1146-9.745-.2485-.822-1.2615-1.551-2.1311-1.551-.2294 0-.3345-.055-.4396-.24-.3631-.664-1.4239-1.236-2.3031-1.236h-.3631v-3.489c0-3.571-.0765-4.605-.4396-5.832-.2962-1.034-3.4116-9.478-3.6601-9.921-.516-.951-1.5768-1.735-2.7331-2.03-2.3508-.591-4.8068.904-5.3037 3.239-.0956.415-.1434.498-.2676.452-.4683-.175-1.8539-.526-2.628-.664-1.2041-.213-3.899-.222-5.084-.01-.7549.139-2.1788.499-2.6471.674-.1242.046-.172-.037-.2675-.452-.5065-2.344-2.9816-3.848-5.342-3.23Zm2.2839 1.606c.9557.461 1.5482 1.319 1.5673 2.242l.0095.545-.8983.498c-1.3761.766-2.3317 1.467-3.5645 2.612-1.7488 1.624-3.0006 3.248-4.0614 5.26-.2771.526-.5065.923-.5065.877 0-.046.086-.434.1816-.868.1911-.812 3.2969-9.413 3.6314-10.031.3058-.582.9652-1.089 1.7297-1.329.4205-.139 1.4525-.037 1.9112.194Zm-9.0785 22.25v1.126l-.4395-.056c-.6594-.064-1.0417-.47-1.0417-1.07 0-.674.4492-1.098 1.1754-1.108l.3058-.009v1.117Zm36.3523-.757c.2484.249.2962.36.2962.748 0 .609-.3822 1.015-1.0416 1.079l-.4396.056v-2.27l.4396.055c.3249.028.5256.12.7454.332Z" />
+        <path d="M70.86 181.196c-2.1693.757-2.8 3.341-1.1946 4.882.6308.609 1.1946.83 2.1502.821.6881 0 .8887-.037 1.2997-.24.7167-.36 1.0225-.646 1.3856-1.292.2867-.526.3154-.655.3249-1.366 0-.923-.2293-1.467-.86-2.076-.7836-.757-2.1215-1.071-3.1058-.729Zm1.5577 1.458c.9843.397 1.2232 1.633.4587 2.372-1.0704 1.043-2.8287.11-2.5516-1.348.172-.867 1.2232-1.384 2.0929-1.024ZM59.2973 186.595c-.2198.406-.3823.756-.3632.775.0765.064 1.2233.6 1.2997.6.0382 0 .2294-.305.4205-.683.3918-.766.4205-.692-.4205-1.154l-.5447-.295-.3918.757ZM65.0866 187.831c-1.8157.443-3.5263 2.012-4.2525 3.885-.2485.628-.4874 1.735-.4874 2.261v.36h1.529v-.35c0-.185.0669-.628.1433-.979.5161-2.334 2.5707-4.014 4.6731-3.802.9365.093 1.6054.471 2.5802 1.449.7645.775.8314.868.774 1.135-.1337.628-.0955.6-.6976.462-1.0703-.249-2.2553-.12-3.1727.332-.1815.092-.8696.526-1.5194.969-1.3952.941-2.3126 1.393-3.0294 1.486l-.516.055v1.486l.5734-.065c1.1563-.129 2.1788-.59 3.832-1.744.5448-.378 1.185-.784 1.4239-.895.5065-.249 1.4717-.341 2.1024-.194.3154.065.4396.065.497-.018.1337-.203.086.064-.7359 4.014-.6116 2.935-.8123 4.079-.8123 4.633v.729l-.8791.84c-.583.544-1.1372.978-1.6055 1.227-1.185.646-2.6758 1.043-3.8894 1.043h-.5065v1.504l.9365-.055c2.3413-.148 4.4915-1.071 6.1638-2.649l.6785-.636.6976.618c.9748.858 1.4622 1.144 2.026 1.2.8314.073 1.1276-.074 2.4177-1.2l.7072-.618.6689.618c1.7488 1.661 4.2717 2.695 6.5461 2.695h.5638V206.178l-.8791-.056c-2.0738-.138-3.8895-.959-5.3229-2.418l-.6785-.692v-.72c0-.526-.2102-1.725-.8123-4.614-.8218-3.95-.8696-4.217-.7358-4.014.0573.083.1815.083.4969.018.6307-.147 1.5959-.055 2.1024.194.2389.111.8792.517 1.4239.895 1.6532 1.154 2.6757 1.615 3.8321 1.744l.5733.065v-1.486l-.516-.055c-.7167-.093-1.6341-.545-3.0294-1.486-1.3665-.923-1.6532-1.08-2.4273-1.292-.6116-.175-2.045-.157-2.5515.037-.2771.101-.2867.101-.4109-.508-.0574-.267.0095-.36.774-1.135.9748-.978 1.6437-1.356 2.5802-1.449 2.112-.212 4.157 1.458 4.6731 3.821.0764.369.1433.821.1433.997v.313h1.5577l-.0573-.673c-.258-2.833-2.2458-5.261-4.8069-5.842-1.2614-.286-2.8-.018-3.9467.692-.2676.166-.7167.545-1.0034.849l-.5256.545H70.6579l-.4682-.499c-1.013-1.061-2.2171-1.615-3.6792-1.679-.5638-.019-1.0512.009-1.4239.101Zm7.9413 8.73c.6881 3.304 1.0703 5.325 1.0799 5.722v.6l-1.013.877c-.5638.48-1.1181.877-1.2423.895-.1816.028-.4778-.175-1.2806-.877l-1.0512-.904v-.554c0-.461 1.9686-10.354 2.1502-10.825.1625-.425.3536.286 1.357 5.066Z" />
+        <path d="M63.2145 198.915c-2.1693.756-2.8 3.34-1.1946 4.882.6307.609 1.1946.83 2.1502.821.6881 0 .8887-.037 1.2997-.24.7167-.36 1.0225-.646 1.3856-1.292.2867-.526.3154-.655.3249-1.366 0-.923-.2293-1.467-.86-2.076-.7836-.757-2.1215-1.071-3.1058-.729Zm1.5576 1.458c.9843.397 1.2233 1.633.4587 2.371-.6211.6-1.5003.6-2.1215 0-1.1849-1.135.1243-2.999 1.6628-2.371ZM78.5055 198.915c-2.1693.756-2.8 3.34-1.1946 4.882.6308.609 1.1946.83 2.1502.821.6881 0 .8888-.037 1.2997-.24.7167-.36 1.0225-.646 1.3856-1.292.2867-.526.3154-.655.325-1.366 0-.923-.2294-1.467-.8601-2.076-.7836-.757-2.1215-1.071-3.1058-.729Zm1.5577 1.458c.9843.397 1.2232 1.633.4587 2.371-.6212.6-1.5004.6-2.1215 0-1.185-1.135.1242-2.999 1.6628-2.371ZM66.6996 207.12c-.2772.258-.2963.48-.086 1.163.86 2.824 4.1283 4.457 6.9379 3.47.9174-.323 1.4907-.674 2.1406-1.32.6211-.618 1.0798-1.366 1.3187-2.15.2103-.683.1912-.905-.086-1.163l-.2293-.231h-9.7666l-.2293.231Zm8.486 1.532c-.3823.637-.9939 1.19-1.7297 1.532-.6881.332-.7741.35-1.6437.35-.8696 0-.9556-.018-1.6437-.35-.7358-.342-1.3474-.895-1.7297-1.532l-.172-.286h7.0908l-.172.286ZM48.8789 216.486v.739h1.529V215.748h-1.529v.738ZM93.2207 216.486v.739h1.529V215.748h-1.529v.738Z" />
+      </g>
+      <rect
+        width="47.9283"
+        height="46.2505"
+        x="47.8496"
+        y="171.212"
+        stroke="red"
+        rx="23.1253"
+      />
+      <path
+        stroke="#C4C4C4"
+        d="M.7891 291.437H438.777M.7891 291.437H438.777M.7891 291.437H438.777M.7891 291.437H438.777M.7891 291.437H438.777"
+      />
+      <path
+        fill="#000"
+        d="M19.4552 263.903v-2.906l7.2869-11.48h2.5057v4.023h-1.483l-4.5937 7.27v.136h10.3551v2.957h-14.071Zm8.3778 3.069v-3.955l.0682-1.287v-12.213h3.4602v17.455H27.833ZM123.45 260.214c1.437 0 2.548.173 3.332.518.784.345 1.176.845 1.176 1.498 0 .663-.392 1.162-1.176 1.498-.784.345-1.895.518-3.332.518-1.428 0-2.534-.173-3.318-.518-.784-.336-1.176-.835-1.176-1.498 0-.653.392-1.153 1.176-1.498.784-.345 1.89-.518 3.318-.518Zm.014 1.344c-.906 0-1.568.051-1.988.154-.42.103-.63.28-.63.532 0 .233.21.406.63.518.42.103 1.082.154 1.988.154.896 0 1.549-.051 1.96-.154.42-.112.63-.285.63-.518 0-.252-.21-.429-.63-.532-.411-.103-1.064-.154-1.96-.154Zm-5.292-9.352h10.57v1.4h-10.57v-1.4Zm5.306 1.736c1.4 0 2.473.159 3.22.476.746.308 1.12.77 1.12 1.386 0 .607-.374 1.069-1.12 1.386-.747.308-1.82.462-3.22.462-1.382 0-2.45-.154-3.206-.462-.756-.317-1.134-.779-1.134-1.386 0-.616.378-1.078 1.134-1.386.756-.317 1.824-.476 3.206-.476Zm0 1.246c-.794 0-1.382.051-1.764.154-.383.093-.574.247-.574.462 0 .205.191.355.574.448.382.093.97.14 1.764.14.812 0 1.404-.047 1.778-.14.382-.093.574-.243.574-.448 0-.215-.192-.369-.574-.462-.374-.103-.966-.154-1.778-.154Zm-.924-4.018h1.848v2.002h-1.848v-2.002Zm-4.942 7.112h11.732v1.442h-11.732v-1.442Zm4.942-1.218h1.848v1.82h-1.848v-1.82Zm16.855-5.796h1.862v6.51h-1.862v-6.51Zm-4.074.49h2.044c0 1.176-.22 2.184-.658 3.024-.439.84-1.106 1.526-2.002 2.058-.887.532-2.016.938-3.388 1.218l-.644-1.498c.868-.159 1.596-.359 2.184-.602.597-.243 1.073-.527 1.428-.854.364-.327.625-.681.784-1.064.168-.383.252-.789.252-1.218v-1.064Zm-3.962 0h5.264v1.47h-5.264v-1.47Zm1.26 6.552h8.638v3.528h-6.762v1.4h-1.862v-2.744h6.748v-.756h-6.762v-1.428Zm.014 4.368h8.946v1.442h-8.946v-1.442Zm10.751-5.306h11.718v1.47h-11.718v-1.47Zm4.928-1.946h1.862v2.52h-1.862v-2.52Zm-3.514-.588h8.974v1.456h-8.974v-1.456Zm0-3.01h8.904v1.47h-7.056v2.464h-1.848v-3.934Zm4.396 7.658c1.4 0 2.501.21 3.304.63.802.411 1.204.999 1.204 1.764 0 .756-.402 1.339-1.204 1.75-.803.42-1.904.63-3.304.63s-2.502-.21-3.304-.63c-.794-.411-1.19-.994-1.19-1.75 0-.765.396-1.353 1.19-1.764.802-.42 1.904-.63 3.304-.63Zm0 1.414c-.579 0-1.064.037-1.456.112-.383.065-.672.173-.868.322-.196.14-.294.317-.294.532 0 .224.098.411.294.56.196.14.485.247.868.322.392.065.877.098 1.456.098.578 0 1.059-.033 1.442-.098.392-.075.686-.182.882-.322.196-.149.294-.336.294-.56 0-.215-.098-.392-.294-.532-.196-.149-.49-.257-.882-.322-.383-.075-.864-.112-1.442-.112ZM225.562 252.818V263h-2.153v-8.138h-.06l-2.331 1.461v-1.909l2.52-1.596h2.024Zm6.412 10.406c-.855-.004-1.591-.214-2.208-.632-.613-.417-1.085-1.022-1.417-1.814-.328-.792-.49-1.745-.487-2.859 0-1.11.164-2.057.492-2.839.332-.782.804-1.377 1.417-1.785.617-.411 1.351-.616 2.203-.616.852 0 1.584.205 2.197.616.617.411 1.091 1.008 1.422 1.79.331.779.496 1.724.492 2.834 0 1.117-.165 2.072-.497 2.864-.328.792-.799 1.397-1.412 1.814-.613.418-1.347.627-2.202.627Zm0-1.785c.583 0 1.049-.293 1.397-.88.348-.587.52-1.467.517-2.64 0-.772-.08-1.415-.239-1.929-.156-.514-.378-.9-.666-1.158-.285-.259-.621-.388-1.009-.388-.58 0-1.044.29-1.392.87-.348.58-.524 1.448-.527 2.605 0 .782.078 1.435.233 1.959.159.52.383.911.672 1.173.288.259.626.388 1.014.388Zm9.638 1.785c-.855-.004-1.591-.214-2.207-.632-.613-.417-1.085-1.022-1.417-1.814-.328-.792-.49-1.745-.487-2.859 0-1.11.164-2.057.492-2.839.332-.782.804-1.377 1.417-1.785.616-.411 1.351-.616 2.202-.616.852 0 1.585.205 2.198.616.616.411 1.09 1.008 1.422 1.79.331.779.495 1.724.492 2.834 0 1.117-.166 2.072-.497 2.864-.328.792-.799 1.397-1.412 1.814-.613.418-1.347.627-2.203.627Zm0-1.785c.584 0 1.049-.293 1.397-.88.348-.587.521-1.467.517-2.64 0-.772-.079-1.415-.238-1.929-.156-.514-.378-.9-.666-1.158-.285-.259-.622-.388-1.01-.388-.58 0-1.044.29-1.392.87-.348.58-.523 1.448-.527 2.605 0 .782.078 1.435.234 1.959.159.52.383.911.671 1.173.289.259.627.388 1.014.388ZM306.053 253.362v10.182H303.9v-8.139h-.059l-2.332 1.462v-1.909l2.521-1.596h2.023ZM353.822 261v-1.551l3.625-3.356c.308-.298.566-.567.775-.805.212-.239.373-.473.482-.701.11-.232.165-.483.165-.751 0-.298-.068-.555-.204-.771-.136-.218-.322-.386-.557-.502-.235-.119-.502-.179-.801-.179-.311 0-.583.063-.815.189-.232.126-.411.307-.537.542-.126.235-.189.515-.189.84h-2.043c0-.666.151-1.244.452-1.735.302-.49.725-.87 1.268-1.138.544-.269 1.17-.403 1.879-.403.73 0 1.364.129 1.904.388.544.255.967.61 1.268 1.064.302.454.453.974.453 1.561 0 .384-.077.764-.229 1.138-.149.375-.416.791-.8 1.248-.385.454-.927.999-1.626 1.636l-1.487 1.456v.07h4.276V261h-7.259Zm12.811.224c-.855-.004-1.591-.214-2.207-.632-.614-.417-1.086-1.022-1.417-1.814-.329-.792-.491-1.745-.488-2.859 0-1.11.164-2.057.493-2.839.331-.782.803-1.377 1.416-1.785.617-.411 1.351-.616 2.203-.616.852 0 1.584.205 2.197.616.617.411 1.091 1.008 1.422 1.79.332.779.496 1.724.492 2.834 0 1.117-.165 2.072-.497 2.864-.328.792-.799 1.397-1.412 1.814-.613.418-1.347.627-2.202.627Zm0-1.785c.583 0 1.049-.293 1.397-.88.348-.587.52-1.467.517-2.64 0-.772-.08-1.415-.239-1.929-.155-.514-.378-.9-.666-1.158-.285-.259-.621-.388-1.009-.388-.58 0-1.044.29-1.392.87-.348.58-.524 1.448-.527 2.605 0 .782.078 1.435.234 1.959.159.52.382.911.671 1.173.288.259.626.388 1.014.388Zm5.646 1.561v-1.551l3.625-3.356c.308-.298.566-.567.775-.805.212-.239.373-.473.482-.701.11-.232.165-.483.165-.751 0-.298-.068-.555-.204-.771-.136-.218-.322-.386-.557-.502-.235-.119-.502-.179-.801-.179-.311 0-.583.063-.815.189-.232.126-.411.307-.537.542-.126.235-.189.515-.189.84h-2.043c0-.666.151-1.244.452-1.735.302-.49.725-.87 1.268-1.138.544-.269 1.17-.403 1.879-.403.73 0 1.364.129 1.905.388.543.255.966.61 1.267 1.064.302.454.453.974.453 1.561 0 .384-.077.764-.229 1.138-.149.375-.416.791-.8 1.248-.385.454-.927.999-1.626 1.636l-1.487 1.456v.07h4.276V261h-7.259Zm12.537.139c-.702 0-1.329-.129-1.879-.388-.547-.258-.981-.614-1.302-1.068-.322-.455-.489-.975-.502-1.562h2.088c.023.395.189.715.497.96s.674.368 1.098.368c.339 0 .637-.075.895-.224.262-.152.466-.363.612-.631.149-.272.224-.584.224-.935 0-.358-.077-.673-.229-.945-.149-.271-.356-.483-.622-.636-.265-.152-.568-.23-.909-.234-.299 0-.589.062-.87.184-.279.123-.496.29-.652.503l-1.914-.343.483-5.37h6.224v1.76h-4.45l-.263 2.551h.06c.179-.252.449-.461.81-.627.361-.166.766-.248 1.213-.248.613 0 1.16.144 1.641.432.48.288.86.685 1.138 1.188.279.501.416 1.077.413 1.73.003.686-.156 1.296-.478 1.83-.318.53-.763.948-1.337 1.253-.57.301-1.233.452-1.989.452Zm6.235-.01c-.328 0-.61-.116-.845-.348-.232-.235-.348-.517-.348-.845 0-.325.116-.603.348-.835.235-.232.517-.348.845-.348.318 0 .596.116.835.348.239.232.358.51.358.835 0 .219-.056.419-.169.602-.109.179-.254.323-.433.432-.178.106-.376.159-.591.159Zm6.494.095c-.855-.004-1.591-.214-2.207-.632-.614-.417-1.086-1.022-1.417-1.814-.328-.792-.491-1.745-.487-2.859 0-1.11.164-2.057.492-2.839.331-.782.803-1.377 1.417-1.785.616-.411 1.35-.616 2.202-.616.852 0 1.584.205 2.197.616.617.411 1.091 1.008 1.422 1.79.332.779.496 1.724.493 2.834 0 1.117-.166 2.072-.498 2.864-.328.792-.798 1.397-1.412 1.814-.613.418-1.347.627-2.202.627Zm0-1.785c.583 0 1.049-.293 1.397-.88.348-.587.52-1.467.517-2.64 0-.772-.079-1.415-.239-1.929-.155-.514-.377-.9-.666-1.158-.285-.259-.621-.388-1.009-.388-.58 0-1.044.29-1.392.87-.348.58-.524 1.448-.527 2.605 0 .782.078 1.435.234 1.959.159.52.382.911.671 1.173.288.259.626.388 1.014.388Zm10.076-8.621V261h-2.153v-8.138h-.059l-2.332 1.461v-1.909l2.521-1.596h2.023Zm3.131 10.311c-.328 0-.61-.116-.845-.348-.232-.235-.348-.517-.348-.845 0-.325.116-.603.348-.835.235-.232.517-.348.845-.348.318 0 .597.116.835.348.239.232.358.51.358.835 0 .219-.056.419-.169.602-.109.179-.253.323-.432.432-.179.106-.377.159-.592.159Zm6.07-10.311V261h-2.152v-8.138h-.06l-2.332 1.461v-1.909l2.521-1.596h2.023ZM419.645 261l4.221-8.352v-.07h-4.917v-1.76h7.144v1.785L421.867 261h-2.222Z"
+      />
+      <g fill="#FDA970" clipPath="url(#d)">
+        <path d="M61.025 234.164c-1.1277.286-2.1884 1.089-2.6949 2.021-.2485.443-3.3638 8.887-3.6601 9.92-.3631 1.228-.4396 2.261-.4396 5.833v3.488h-.3631c-.8792 0-1.9399.573-2.3031 1.237-.1051.175-.2102.24-.4204.24-.8506 0-1.9113.757-2.1502 1.532-.086.286-.1147 2.667-.1147 9.764v9.376h1.529V258.832l.2389-.231c.3441-.332.5256-.258.6881.277.2867.932 1.4334 1.716 2.5324 1.716h.3631v2.105c0 2.334.0765 2.787.5925 3.728.4014.729 1.3093 1.541 2.112 1.873.6307.268.6403.277.7932.785.086.277.344.987.5829 1.569 1.099 2.704 2.7235 5.112 4.6348 6.884 7.0621 6.543 16.7427 4.218 21.6068-5.186.2389-.453.4492-.905.4683-.988.0191-.129-.1338-.231-.6307-.434-.3632-.147-.669-.267-.6881-.249-.0095.009-.2962.554-.6402 1.218-2.284 4.356-5.6861 7.079-9.6615 7.752-2.4655.416-5.2273-.193-7.4635-1.642-3.985-2.584-6.7658-7.411-7.6546-13.308-.2389-1.56-.2389-5.205-.0095-6.691.2389-1.467.602-3.008.9843-4.162.1815-.535.3249-1.015.3249-1.061 0-.046-.3058-.194-.6785-.332l-.6785-.24-.1338.323c-.4396 1.006-1.1563 3.931-1.3761 5.565-.2867 2.187-.2103 5.352.1816 7.567.0955.535.1624.988.1433 1.006-.0764.083-.7836-.729-1.0034-1.154l-.2294-.452v-5.629c0-5.99.0191-6.303.4874-8.112.8505-3.313 2.6758-6.359 5.1413-8.61 2.3222-2.123 4.7686-3.397 7.7502-4.015 1.3761-.286 3.8894-.286 5.2655 0 1.7297.36 3.5167 1.043 4.7973 1.827.2771.176.516.314.5351.314.0191 0 .2198-.277.4492-.609.3345-.498.3822-.618.2676-.692a37.4413 37.4413 0 0 0-.9843-.563l-.841-.48.0478-.517c.086-.932.6594-1.744 1.5481-2.196.6594-.342 1.7966-.351 2.4369-.028.5542.277 1.0225.702 1.2518 1.145.1052.193.927 2.353 1.8157 4.798 1.4144 3.867 1.9878 5.63 1.9973 6.156 0 .074-.2484-.342-.5543-.923-.6784-1.301-1.3761-2.344-2.3795-3.553-.8123-.969-2.4655-2.575-2.5993-2.519-.0478.018-.2771.258-.516.535l-.4396.507 1.1276 1.108c.6116.609 1.3188 1.356 1.5577 1.661 2.1406 2.741 3.4498 5.925 3.8034 9.321.0669.664.0956 2.778.0765 6.045l-.0287 5.029-.3249.582c-.2772.489-.8314 1.107-.9174 1.015-.0096-.019.0573-.471.1529-1.006 1.3187-7.457-.8983-15.375-5.6287-20.137-3.6696-3.673-8.4669-4.983-12.9679-3.535-1.9495.637-3.4881 1.56-5.1509 3.101-1.1276 1.052-2.3317 2.464-2.1979 2.575.2771.231 1.0798.748 1.1467.738.0478 0 .4301-.424.8601-.941 2.7809-3.396 6.7468-5.057 10.598-4.448 3.2969.517 6.2498 2.538 8.5146 5.814 3.4594 5.029 4.568 12.2 2.8669 18.642-.1338.507-.3058 1.098-.3822 1.31-.086.231-.1051.415-.0574.443.1529.092 1.1659.415 1.2901.415.0669 0 .1912-.24.2772-.535.1529-.526.1529-.526.7836-.793 1.2805-.545 2.3604-1.855 2.6184-3.184.0574-.268.0956-1.403.0956-2.511v-2.021h.3631c1.099 0 2.2458-.784 2.5325-1.716.1624-.535.344-.609.688-.277l.2389.231V277.575h1.529v-9.376c0-6.986-.0286-9.478-.1146-9.746-.2485-.821-1.2615-1.55-2.1311-1.55-.2294 0-.3345-.055-.4396-.24-.3631-.664-1.4239-1.237-2.3031-1.237h-.3631v-3.488c0-3.572-.0765-4.605-.4396-5.833-.2962-1.033-3.4116-9.477-3.6601-9.92-.516-.951-1.5768-1.735-2.7331-2.031-2.3508-.59-4.8068.905-5.3037 3.24-.0956.415-.1434.498-.2676.452-.4683-.176-1.8539-.526-2.628-.665-1.2041-.212-3.899-.221-5.084-.009-.7549.139-2.1788.498-2.6471.674-.1242.046-.172-.037-.2675-.452-.5065-2.344-2.9816-3.849-5.342-3.23Zm2.2839 1.605c.9557.462 1.5482 1.32 1.5673 2.243l.0095.544-.8983.499c-1.3761.766-2.3317 1.467-3.5645 2.611-1.7488 1.625-3.0006 3.249-4.0614 5.261-.2771.526-.5065.923-.5065.877 0-.047.086-.434.1816-.868.1911-.812 3.2969-9.413 3.6314-10.031.3058-.582.9652-1.089 1.7297-1.329.4205-.139 1.4525-.037 1.9112.193Zm-9.0785 22.251v1.125l-.4395-.055c-.6594-.064-1.0417-.471-1.0417-1.07 0-.674.4492-1.099 1.1754-1.108l.3058-.009v1.117Zm36.3523-.757c.2484.249.2962.36.2962.747 0 .609-.3822 1.016-1.0416 1.08l-.4396.055v-2.27l.4396.056c.3249.027.5256.12.7454.332Z" />
+        <path d="M70.86 244.5c-2.1693.756-2.8 3.34-1.1946 4.882.6308.609 1.1946.83 2.1502.821.6881 0 .8887-.037 1.2997-.24.7167-.36 1.0225-.646 1.3856-1.292.2867-.526.3154-.655.3249-1.366 0-.923-.2293-1.467-.86-2.076-.7836-.757-2.1215-1.071-3.1058-.729Zm1.5577 1.458c.9843.397 1.2232 1.633.4587 2.371-1.0704 1.043-2.8287.111-2.5516-1.347.172-.867 1.2232-1.384 2.0929-1.024ZM59.2973 249.898c-.2198.406-.3823.757-.3632.776.0765.064 1.2233.599 1.2997.599.0382 0 .2294-.304.4205-.682.3918-.766.4205-.693-.4205-1.154l-.5447-.295-.3918.756ZM65.0866 251.135c-1.8157.443-3.5263 2.012-4.2525 3.885-.2485.628-.4874 1.735-.4874 2.261v.36h1.529v-.351c0-.184.0669-.627.1433-.978.5161-2.335 2.5707-4.014 4.6731-3.802.9365.092 1.6054.471 2.5802 1.449.7645.775.8314.867.774 1.135-.1337.628-.0955.6-.6976.461-1.0703-.249-2.2553-.12-3.1727.333-.1815.092-.8696.526-1.5194.969-1.3952.941-2.3126 1.393-3.0294 1.485l-.516.056v1.486l.5734-.065c1.1563-.129 2.1788-.591 3.832-1.744.5448-.379 1.185-.785 1.4239-.895.5065-.249 1.4717-.342 2.1024-.194.3154.064.4396.064.497-.019.1337-.203.086.065-.7359 4.015-.6116 2.935-.8123 4.079-.8123 4.633v.729l-.8791.84c-.583.544-1.1372.978-1.6055 1.227-1.185.646-2.6758 1.043-3.8894 1.043h-.5065v1.504l.9365-.055c2.3413-.148 4.4915-1.071 6.1638-2.649l.6785-.637.6976.619c.9748.858 1.4622 1.144 2.026 1.199.8314.074 1.1276-.074 2.4177-1.199l.7072-.619.6689.619c1.7488 1.661 4.2717 2.694 6.5461 2.694h.5638V269.481l-.8791-.055c-2.0738-.138-3.8895-.96-5.3229-2.418l-.6785-.692v-.72c0-.526-.2102-1.726-.8123-4.614-.8218-3.95-.8696-4.218-.7358-4.015.0573.083.1815.083.4969.019.6307-.148 1.5959-.055 2.1024.194.2389.11.8792.516 1.4239.895 1.6532 1.153 2.6757 1.615 3.8321 1.744l.5733.065v-1.486l-.516-.056c-.7167-.092-1.6341-.544-3.0294-1.485-1.3665-.923-1.6532-1.08-2.4273-1.292-.6116-.176-2.045-.157-2.5515.037-.2771.101-.2867.101-.4109-.508-.0574-.268.0095-.36.774-1.135.9748-.978 1.6437-1.357 2.5802-1.449 2.112-.212 4.157 1.458 4.6731 3.821.0764.369.1433.821.1433.996v.314h1.5577l-.0573-.674c-.258-2.833-2.2458-5.26-4.8069-5.841-1.2614-.286-2.8-.019-3.9467.692-.2676.166-.7167.544-1.0034.849l-.5256.544H70.6579l-.4682-.498c-1.013-1.061-2.2171-1.615-3.6792-1.68-.5638-.018-1.0512.01-1.4239.102Zm7.9413 8.73c.6881 3.304 1.0703 5.325 1.0799 5.722v.6l-1.013.877c-.5638.479-1.1181.876-1.2423.895-.1816.027-.4778-.176-1.2806-.877l-1.0512-.904v-.554c0-.462 1.9686-10.355 2.1502-10.825.1625-.425.3536.286 1.357 5.066Z" />
+        <path d="M63.2145 262.218c-2.1693.757-2.8 3.341-1.1946 4.882.6307.609 1.1946.831 2.1502.822.6881 0 .8887-.037 1.2997-.24.7167-.36 1.0225-.646 1.3856-1.292.2867-.526.3154-.656.3249-1.366 0-.923-.2293-1.468-.86-2.077-.7836-.756-2.1215-1.07-3.1058-.729Zm1.5576 1.458c.9843.397 1.2233 1.634.4587 2.372-.6211.6-1.5003.6-2.1215 0-1.1849-1.135.1243-2.999 1.6628-2.372ZM78.5055 262.218c-2.1693.757-2.8 3.341-1.1946 4.882.6308.609 1.1946.831 2.1502.822.6881 0 .8888-.037 1.2997-.24.7167-.36 1.0225-.646 1.3856-1.292.2867-.526.3154-.656.325-1.366 0-.923-.2294-1.468-.8601-2.077-.7836-.756-2.1215-1.07-3.1058-.729Zm1.5577 1.458c.9843.397 1.2232 1.634.4587 2.372-.6212.6-1.5004.6-2.1215 0-1.185-1.135.1242-2.999 1.6628-2.372ZM66.6996 270.424c-.2772.258-.2963.479-.086 1.162.86 2.824 4.1283 4.458 6.9379 3.47.9174-.323 1.4907-.673 2.1406-1.319.6211-.619 1.0798-1.366 1.3187-2.151.2103-.683.1912-.904-.086-1.162l-.2293-.231h-9.7666l-.2293.231Zm8.486 1.531c-.3823.637-.9939 1.191-1.7297 1.532-.6881.333-.7741.351-1.6437.351-.8696 0-.9556-.018-1.6437-.351-.7358-.341-1.3474-.895-1.7297-1.532l-.172-.286h7.0908l-.172.286ZM48.8789 279.79v.738h1.529v-1.476h-1.529v.738ZM93.2207 279.79v.738h1.529v-1.476h-1.529v.738Z" />
+      </g>
+      <rect
+        width="47.9283"
+        height="46.2505"
+        x="47.8496"
+        y="234.516"
+        stroke="#FDA970"
+        rx="23.1253"
+      />
+      <path
+        fill="#000"
+        d="M229.579 11.25h2.45v1.512h-2.45V11.25Zm1.96-2.968h1.876v7.49h-1.876v-7.49Zm-6.776 8.008h8.652v4.816h-8.652V16.29Zm6.818 1.456h-4.984v1.89h4.984v-1.89Zm-6.02-8.274h1.512v.826c0 .812-.13 1.5867-.392 2.324-.261.7373-.662 1.386-1.204 1.946-.532.5507-1.218.9613-2.058 1.232l-.924-1.47c.532-.1773.99-.406 1.372-.686.392-.2893.71-.616.952-.98.252-.364.439-.7467.56-1.148a4.165 4.165 0 0 0 .182-1.218v-.826Zm.392 0h1.484v.812c0 .504.103 1.008.308 1.512.206.4947.523.9427.952 1.344.439.4013.999.7187 1.68.952l-.91 1.456c-.802-.2707-1.465-.672-1.988-1.204-.513-.532-.896-1.148-1.148-1.848-.252-.7-.378-1.4373-.378-2.212v-.812Zm-2.954-.546h6.972v1.484h-6.972V8.926Zm17.331-.28h1.624v.616c0 .4947-.079.9753-.238 1.442-.149.4573-.373.8867-.672 1.288-.299.4013-.663.7607-1.092 1.078-.429.3173-.929.588-1.498.812-.56.2147-1.176.378-1.848.49l-.728-1.498c.597-.0747 1.129-.2007 1.596-.378.476-.1773.891-.3873 1.246-.63.364-.2427.663-.5087.896-.798.233-.2893.411-.588.532-.896.121-.308.182-.6113.182-.91v-.616Zm.35 0h1.624v.616c0 .2987.061.602.182.91.121.308.299.6067.532.896.233.2893.527.5553.882.798.364.2427.779.4527 1.246.63.476.1773 1.013.3033 1.61.378l-.728 1.498c-.681-.112-1.302-.2753-1.862-.49-.56-.224-1.055-.4947-1.484-.812-.429-.3267-.793-.6907-1.092-1.092-.299-.4013-.527-.8307-.686-1.288-.149-.4573-.224-.9333-.224-1.428v-.616Zm-.35 7.84h1.862v4.76h-1.862v-4.76Zm-4.886-1.176h11.732v1.512h-11.732V15.31ZM286.336 9.276h1.526v1.876c0 .84-.074 1.6473-.224 2.422-.149.7653-.378 1.484-.686 2.156-.298.6627-.686 1.2507-1.162 1.764-.466.504-1.022.9053-1.666 1.204l-1.106-1.54c.57-.252 1.06-.5833 1.47-.994.42-.42.766-.896 1.036-1.428.271-.5413.472-1.1153.602-1.722.14-.6067.21-1.2273.21-1.862V9.276Zm.364 0h1.512v1.876c0 .6067.066 1.1993.196 1.778.131.5787.327 1.1247.588 1.638.271.5133.607.9707 1.008 1.372.411.392.892.7047 1.442.938l-1.078 1.512c-.625-.2893-1.171-.6767-1.638-1.162-.457-.4853-.835-1.0453-1.134-1.68-.298-.644-.522-1.3347-.672-2.072-.149-.7467-.224-1.5213-.224-2.324V9.276Zm5.376-1.022h1.862V21.26h-1.862V8.254Zm5.361 5.796h9.114v1.498h-9.114V14.05Zm-1.372 4.214h11.746v1.526h-11.746v-1.526Zm4.928-3.528h1.848v4.018h-1.848v-4.018Zm-3.556-5.572h9.016v1.484h-7.168v4.046h-1.848v-5.53Zm20.439-.882h1.862V17.9h-1.862V8.282Zm-8.498.966h7.602v1.33h-7.602v-1.33Zm3.794 1.666c.653 0 1.218.0793 1.694.238.485.1493.863.364 1.134.644.271.28.406.6113.406.994 0 .3827-.135.7187-.406 1.008s-.649.5133-1.134.672c-.476.1493-1.041.224-1.694.224-.644 0-1.209-.0747-1.694-.224-.476-.1587-.849-.3827-1.12-.672-.261-.2893-.392-.6253-.392-1.008 0-.3827.131-.714.392-.994.271-.28.644-.4947 1.12-.644.485-.1587 1.05-.238 1.694-.238Zm0 1.218c-.439 0-.789.056-1.05.168-.261.1027-.392.266-.392.49 0 .224.131.3967.392.518.261.112.611.168 1.05.168.448 0 .803-.056 1.064-.168.261-.1213.392-.294.392-.518 0-.224-.131-.3873-.392-.49-.261-.112-.616-.168-1.064-.168Zm-.924-3.948h1.876v1.694h-1.876V8.184Zm0 5.88h1.876v1.792h-1.876v-1.792Zm-3.024 2.744-.21-1.414c.784 0 1.629-.0047 2.534-.014.915-.0093 1.848-.042 2.8-.098.961-.056 1.881-.1353 2.758-.238l.14 1.246a30.281 30.281 0 0 1-2.786.35c-.943.0747-1.862.1213-2.758.14-.887.0187-1.713.028-2.478.028Zm5.068.434h1.596v.154c0 .4947-.112.9613-.336 1.4-.224.4293-.546.812-.966 1.148-.42.3267-.947.5973-1.582.812-.625.2147-1.353.364-2.184.448l-.616-1.4c.728-.056 1.349-.1633 1.862-.322.513-.168.933-.3593 1.26-.574.327-.2147.569-.4527.728-.714.159-.2707.238-.5367.238-.798v-.154Zm.294 0h1.61v.154c0 .2613.079.5273.238.798.159.2613.401.4993.728.714.327.2147.747.406 1.26.574.513.1587 1.129.266 1.848.322l-.602 1.4c-.831-.084-1.563-.2333-2.198-.448-.625-.2147-1.153-.4853-1.582-.812-.42-.336-.742-.7187-.966-1.148-.224-.4387-.336-.9053-.336-1.4v-.154Zm12.123-8.596h1.624v.616c0 .4947-.079.9753-.238 1.442-.149.4573-.373.8867-.672 1.288-.299.4013-.663.7607-1.092 1.078-.429.3173-.929.588-1.498.812-.56.2147-1.176.378-1.848.49l-.728-1.498c.597-.0747 1.129-.2007 1.596-.378.476-.1773.891-.3873 1.246-.63.364-.2427.663-.5087.896-.798.233-.2893.411-.588.532-.896.121-.308.182-.6113.182-.91v-.616Zm.35 0h1.624v.616c0 .2987.061.602.182.91.121.308.299.6067.532.896.233.2893.527.5553.882.798.364.2427.779.4527 1.246.63.476.1773 1.013.3033 1.61.378l-.728 1.498c-.681-.112-1.302-.2753-1.862-.49-.56-.224-1.055-.4947-1.484-.812-.429-.3267-.793-.6907-1.092-1.092-.299-.4013-.527-.8307-.686-1.288-.149-.4573-.224-.9333-.224-1.428v-.616Zm-.35 7.84h1.862v4.76h-1.862v-4.76Zm-4.886-1.176h11.732v1.512h-11.732V15.31ZM135.917 7.795h1.862v13.006h-1.862V7.795Zm-5.124.91c.663 0 1.251.1913 1.764.574.514.3734.915.91 1.204 1.61.299.6907.448 1.5073.448 2.45 0 .952-.149 1.778-.448 2.478-.289.6907-.69 1.2273-1.204 1.61-.513.3733-1.101.56-1.764.56-.653 0-1.236-.1867-1.75-.56-.513-.3827-.919-.9193-1.218-1.61-.289-.7-.434-1.526-.434-2.478 0-.9427.145-1.7593.434-2.45.299-.7.705-1.2366 1.218-1.61.514-.3826 1.097-.574 1.75-.574Zm0 1.694c-.317 0-.597.112-.84.336-.242.224-.434.5553-.574.994-.13.4293-.196.966-.196 1.61 0 .6347.066 1.176.196 1.624.14.4387.332.77.574.994.243.224.523.336.84.336.327 0 .612-.112.854-.336.243-.224.43-.5553.56-.994.14-.448.21-.9893.21-1.624 0-.644-.07-1.1807-.21-1.61-.13-.4387-.317-.77-.56-.994-.242-.224-.527-.336-.854-.336Zm9.113 3.962h11.732v1.47h-11.732v-1.47Zm1.428-6.3h8.876v3.458h-7.014v1.414h-1.834v-2.73h7.014v-.728h-7.042V8.061Zm.028 4.186h9.1v1.414h-9.1v-1.414Zm.014 4.354h8.904v4.046h-8.904v-4.046Zm7.07 1.456h-5.236v1.12h5.236v-1.12ZM363.2 14.204h11.718v1.512H363.2v-1.512Zm1.414-2.366h8.974v1.47h-8.974v-1.47Zm0-3.22h8.904v1.498h-7.056v2.45h-1.848V8.618Zm4.396 7.854c1.4 0 2.502.21 3.304.63.803.4107 1.204.9987 1.204 1.764 0 .756-.401 1.3393-1.204 1.75-.802.42-1.904.63-3.304.63s-2.501-.21-3.304-.63c-.793-.4107-1.19-.994-1.19-1.75 0-.7653.397-1.3533 1.19-1.764.803-.42 1.904-.63 3.304-.63Zm0 1.414c-.578 0-1.064.0373-1.456.112-.382.0653-.672.1727-.868.322-.196.14-.294.322-.294.546 0 .2147.098.3967.294.546.196.14.486.2473.868.322.392.0653.878.098 1.456.098.579 0 1.06-.0327 1.442-.098.392-.0747.686-.182.882-.322.196-.1493.294-.3313.294-.546 0-.224-.098-.406-.294-.546-.196-.1493-.49-.2567-.882-.322-.382-.0747-.863-.112-1.442-.112Zm7.055-2.66h11.732v1.47h-11.732v-1.47Zm1.428-6.706h8.876v3.514h-7.014v1.4h-1.834v-2.716h7.014v-.784h-7.042V8.52Zm.028 4.368h9.1v1.414h-9.1v-1.414Zm3.486.686h1.848v2.45h-1.848v-2.45Zm-3.654 3.85h9.1v3.78h-1.862v-2.296h-7.238v-1.484Zm15.231-8.778c.672 0 1.265.1307 1.778.392.523.252.933.602 1.232 1.05.308.4387.462.9473.462 1.526s-.154 1.092-.462 1.54c-.299.4387-.709.784-1.232 1.036-.513.252-1.106.378-1.778.378-.653 0-1.241-.126-1.764-.378-.523-.252-.938-.5973-1.246-1.036-.299-.448-.448-.9613-.448-1.54s.149-1.092.448-1.54c.308-.448.723-.798 1.246-1.05s1.111-.378 1.764-.378Zm.014 1.512c-.317 0-.602.0607-.854.182-.252.112-.453.2753-.602.49-.14.2147-.21.476-.21.784 0 .308.07.574.21.798.149.2147.345.378.588.49.252.112.541.168.868.168.317 0 .597-.056.84-.168.252-.112.448-.2753.588-.49.149-.224.224-.49.224-.798 0-.308-.075-.5693-.224-.784-.14-.2147-.336-.378-.588-.49-.252-.1213-.532-.182-.84-.182Zm5.25-1.876h1.862v6.482h-1.862V8.282Zm-6.776 7.014h8.638v3.542h-6.79v1.582h-1.834v-2.926h6.776v-.77h-6.79v-1.428Zm.014 4.354h8.946v1.456h-8.946V19.65Zm13.635-9.324h1.47v1.456c0 .7187-.079 1.4373-.238 2.156-.149.7187-.373 1.3953-.672 2.03-.299.6253-.677 1.1853-1.134 1.68-.448.4947-.971.882-1.568 1.162l-1.05-1.47c.541-.252 1.008-.5833 1.4-.994.401-.4107.733-.868.994-1.372.271-.504.471-1.0267.602-1.568.131-.5507.196-1.092.196-1.624v-1.456Zm.392 0h1.47v1.456c0 .4853.065.9847.196 1.498a6.705 6.705 0 0 0 .574 1.484c.261.476.588.91.98 1.302.401.3827.868.6953 1.4.938l-1.036 1.484c-.597-.28-1.12-.6533-1.568-1.12-.439-.476-.812-1.0127-1.12-1.61a8.4689 8.4689 0 0 1-.672-1.918 9.6061 9.6061 0 0 1-.224-2.058v-1.456Zm-3.08-.812h7.126v1.54h-7.126v-1.54Zm8.064-1.232h1.876v12.964h-1.876V8.282Zm1.47 4.942h2.282v1.526h-2.282v-1.526ZM64.1088 13.745h11.718v1.512h-11.718v-1.512Zm1.414-2.366h8.974v1.47h-8.974v-1.47Zm0-3.22h8.904v1.498h-7.056v2.45h-1.848V8.159Zm4.396 7.854c1.4 0 2.5013.21 3.304.63.8026.4107 1.204.9987 1.204 1.764 0 .756-.4014 1.3393-1.204 1.75-.8027.42-1.904.63-3.304.63s-2.5014-.21-3.304-.63c-.7934-.4107-1.19-.994-1.19-1.75 0-.7653.3966-1.3533 1.19-1.764.8026-.42 1.904-.63 3.304-.63Zm0 1.414c-.5787 0-1.064.0373-1.456.112-.3827.0653-.672.1727-.868.322-.196.14-.294.322-.294.546 0 .2147.098.3967.294.546.196.14.4853.2473.868.322.392.0653.8773.098 1.456.098.5786 0 1.0593-.0327 1.442-.098.392-.0747.686-.182.882-.322.196-.1493.294-.3313.294-.546 0-.224-.098-.406-.294-.546-.196-.1493-.49-.2567-.882-.322-.3827-.0747-.8634-.112-1.442-.112Zm8.4409-2.226h1.834v1.344h5.25v-1.344h1.848v5.446h-8.932v-5.446Zm1.834 2.786v1.176h5.25v-1.176h-5.25Zm-1.778-9.66h8.204v1.484h-8.204V8.327Zm-1.442 4.536h11.746v1.498h-11.746v-1.498Zm8.526-4.536h1.834v1.036c0 .5414-.014 1.1153-.042 1.722-.028.5973-.1307 1.302-.308 2.114l-1.876.126c.196-.84.308-1.5727.336-2.198.0373-.6253.056-1.2133.056-1.764V8.327ZM28.8332 8.268h1.778v7.924h-1.778V8.268Zm-1.54 3.178h2.058v1.484h-2.058v-1.484Zm-1.218-2.94h1.736v7.266h-1.736V8.506Zm-5.992 5.586h.896a89.4 89.4 0 0 0 1.666-.014c.504-.0093.98-.0327 1.428-.07.4573-.0467.924-.112 1.4-.196l.14 1.456a20.121 20.121 0 0 1-1.428.196 33.8606 33.8606 0 0 1-1.484.084 95.8284 95.8284 0 0 1-1.722.014h-.896v-1.47Zm-.028-4.942h4.928v3.836h-3.108v2.282h-1.792v-3.64h3.094v-1.022h-3.122V9.15Zm6.16 7.196c.9053 0 1.6893.098 2.352.294.672.196 1.19.476 1.554.84.364.364.546.798.546 1.302 0 .5133-.182.952-.546 1.316-.364.3733-.882.658-1.554.854-.6627.196-1.4467.294-2.352.294-.9053 0-1.694-.098-2.366-.294-.672-.196-1.19-.4807-1.554-.854-.364-.364-.546-.8027-.546-1.316 0-.504.182-.938.546-1.302.364-.364.882-.644 1.554-.84.672-.196 1.4607-.294 2.366-.294Zm0 1.4c-.84 0-1.4887.0887-1.946.266-.448.168-.672.4247-.672.77 0 .3547.224.6207.672.798.4573.168 1.106.252 1.946.252.5507 0 1.0173-.0373 1.4-.112.392-.084.686-.2007.882-.35.2053-.1587.308-.3547.308-.588 0-.224-.1027-.4153-.308-.574-.196-.1587-.49-.2753-.882-.35-.3827-.0747-.8493-.112-1.4-.112Zm11.2409-8.708h1.932c0 1.204-.2006 2.2727-.602 3.206-.4013.9333-1.0453 1.736-1.932 2.408-.8773.6627-2.044 1.1993-3.5 1.61l-.7-1.456c.9147-.252 1.68-.546 2.296-.882.616-.336 1.106-.714 1.47-1.134.364-.42.6254-.8727.784-1.358.168-.4947.252-1.0173.252-1.568v-.826Zm-4.144 0h5.236v1.484h-5.236V9.038Zm4.214 2.422v1.274l-4.62.336-.252-1.442 4.872-.168Zm3.822-3.178h1.862v7.63h-1.862v-7.63Zm-2.52 7.854c.9147 0 1.7034.1027 2.366.308.6627.196 1.176.4853 1.54.868.364.3827.546.84.546 1.372 0 .5413-.182 1.0033-.546 1.386-.364.3827-.8773.672-1.54.868-.6626.2053-1.4513.308-2.366.308-.9053 0-1.6893-.1027-2.352-.308-.6626-.196-1.176-.4853-1.54-.868-.3546-.3827-.532-.8447-.532-1.386 0-.532.1774-.9893.532-1.372.364-.3827.8774-.672 1.54-.868.6627-.2053 1.4467-.308 2.352-.308Zm0 1.442c-.5506 0-1.022.042-1.414.126-.3826.0747-.672.196-.868.364-.196.1587-.294.364-.294.616s.098.462.294.63c.196.168.4854.294.868.378.392.084.8634.126 1.414.126.5694 0 1.0407-.042 1.414-.126.3827-.084.672-.21.868-.378.2054-.168.308-.378.308-.63s-.1026-.4573-.308-.616c-.196-.168-.4853-.2893-.868-.364-.3733-.084-.8446-.126-1.414-.126Z"
+      />
+      <path
+        stroke="#C4C4C4"
+        d="M0 28.459h437.988M0 28.459h437.988M0 28.459h437.988M0 28.459h437.988M0 28.459h437.988"
+      />
+      <defs>
+        <clipPath id="a">
+          <rect
+            width="48.9283"
+            height="47.2505"
+            x="47.3496"
+            y="39.6289"
+            fill="#fff"
+            rx="23.6253"
+          />
+        </clipPath>
+        <clipPath id="b">
+          <rect
+            width="48.9283"
+            height="47.2505"
+            x="47.3496"
+            y="105.17"
+            fill="#fff"
+            rx="23.6253"
+          />
+        </clipPath>
+        <clipPath id="c">
+          <rect
+            width="48.9283"
+            height="47.2505"
+            x="47.3496"
+            y="170.712"
+            fill="#fff"
+            rx="23.6253"
+          />
+        </clipPath>
+        <clipPath id="d">
+          <rect
+            width="48.9283"
+            height="47.2505"
+            x="47.3496"
+            y="234.016"
+            fill="#fff"
+            rx="23.6253"
+          />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
